@@ -1,17 +1,16 @@
 name: flutter_reactive_ble_example
 description: Demonstrates how to use the flutter_reactive_ble plugin.
-version: 5.1.1
+version: 5.3.1
 publish_to: 'none'
 
 environment:
   sdk: '>=2.17.0 <3.0.0'
-  flutter: ">=1.10.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_lints: ^1.0.4
-  flutter_reactive_ble: ^5.1.1
+  flutter_reactive_ble: ^5.3.1
   functional_data: ^1.0.0
   intl: ^0.17.0
 
@@ -20,9 +19,18 @@ dependencies:
 dev_dependencies:
   build_runner: ^2.3.3
   dependency_validator: ^3.1.0
+  flutter_lints: ^1.0.4
   flutter_test:
     sdk: flutter
   functional_data_generator: ^1.1.2
+
+dependency_overrides:
+  flutter_reactive_ble:
+    path: ../packages/flutter_reactive_ble
+  reactive_ble_mobile:
+    path: ../packages/reactive_ble_mobile
+  reactive_ble_platform_interface:
+    path: ../packages/reactive_ble_platform_interface
 
 flutter:
   uses-material-design: true
