@@ -8,29 +8,28 @@ import 'package:provider/provider.dart';
 import 'characteristic_interaction_dialog.dart';
 
 part 'device_interaction_tab.g.dart';
-//ignore_for_file: annotate_overrides
+
+// ignore_for_file: annotate_overrides
 
 class DeviceInteractionTab extends StatelessWidget {
-  final DiscoveredDevice device;
-
   const DeviceInteractionTab({
     required this.device,
     Key? key,
   }) : super(key: key);
 
+  final DiscoveredDevice device;
+
   @override
-  Widget build(BuildContext context) =>
-      Consumer3<BleDeviceConnector, ConnectionStateUpdate, BleDeviceInteractor>(
-        builder: (_, deviceConnector, connectionStateUpdate, serviceDiscoverer,
-                __) =>
-            _DeviceInteractionTab(
+  Widget build(BuildContext context) => Consumer3<BleDeviceConnector, ConnectionStateUpdate, BleDeviceInteractor>(
+        builder: (_, deviceConnector, connectionStateUpdate, serviceDiscoverer, __) => _DeviceInteractionTab(
           viewModel: DeviceInteractionViewModel(
-              deviceId: device.id,
-              connectableStatus: device.connectable,
-              connectionStatus: connectionStateUpdate.connectionState,
-              deviceConnector: deviceConnector,
-              discoverServices: () =>
-                  serviceDiscoverer.discoverServices(device.id)),
+            deviceId: device.id,
+            connectableStatus: device.connectable,
+            connectionStatus: connectionStateUpdate.connectionState,
+            deviceConnector: deviceConnector,
+            discoverServices: () => serviceDiscoverer.discoverServices(device.id),
+            readRssi: () => serviceDiscoverer.readRssi(device.id),
+          ),
         ),
       );
 }
@@ -44,17 +43,19 @@ class DeviceInteractionViewModel extends $DeviceInteractionViewModel {
     required this.connectionStatus,
     required this.deviceConnector,
     required this.discoverServices,
+    required this.readRssi,
   });
 
   final String deviceId;
   final Connectable connectableStatus;
   final DeviceConnectionState connectionStatus;
   final BleDeviceConnector deviceConnector;
-  @CustomEquality(Ignore())
-  final Future<List<DiscoveredService>> Function() discoverServices;
+  final Future<int> Function() readRssi;
 
-  bool get deviceConnected =>
-      connectionStatus == DeviceConnectionState.connected;
+  @CustomEquality(Ignore())
+  final Future<List<Service>> Function() discoverServices;
+
+  bool get deviceConnected => connectionStatus == DeviceConnectionState.connected;
 
   void connect() {
     deviceConnector.connect(deviceId);
@@ -78,7 +79,9 @@ class _DeviceInteractionTab extends StatefulWidget {
 }
 
 class _DeviceInteractionTabState extends State<_DeviceInteractionTab> {
-  late List<DiscoveredService> discoveredServices;
+  late List<Service> discoveredServices;
+
+  int _rssi = 0;
 
   @override
   void initState() {
@@ -93,6 +96,13 @@ class _DeviceInteractionTabState extends State<_DeviceInteractionTab> {
     });
   }
 
+  Future<void> readRssi() async {
+    final rssi = await widget.viewModel.readRssi();
+    setState(() {
+      _rssi = rssi;
+    });
+  }
+
   @override
   Widget build(BuildContext context) => CustomScrollView(
         slivers: [
@@ -100,8 +110,7 @@ class _DeviceInteractionTabState extends State<_DeviceInteractionTab> {
             delegate: SliverChildListDelegate.fixed(
               [
                 Padding(
-                  padding: const EdgeInsetsDirectional.only(
-                      top: 8.0, bottom: 16.0, start: 16.0),
+                  padding: const EdgeInsetsDirectional.only(top: 8.0, bottom: 16.0, start: 16.0),
                   child: Text(
                     "ID: ${widget.viewModel.deviceId}",
                     style: const TextStyle(fontWeight: FontWeight.bold),
@@ -122,27 +131,34 @@ class _DeviceInteractionTabState extends State<_DeviceInteractionTab> {
                   ),
                 ),
                 Padding(
+                  padding: const EdgeInsetsDirectional.only(start: 16.0),
+                  child: Text(
+                    "Rssi: $_rssi dB",
+                    style: const TextStyle(fontWeight: FontWeight.bold),
+                  ),
+                ),
+                Padding(
                   padding: const EdgeInsets.only(top: 16.0),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                  child: Wrap(
+                    alignment: WrapAlignment.spaceEvenly,
                     children: <Widget>[
                       ElevatedButton(
-                        onPressed: !widget.viewModel.deviceConnected
-                            ? widget.viewModel.connect
-                            : null,
+                        onPressed: !widget.viewModel.deviceConnected ? widget.viewModel.connect : null,
                         child: const Text("Connect"),
                       ),
                       ElevatedButton(
-                        onPressed: widget.viewModel.deviceConnected
-                            ? widget.viewModel.disconnect
-                            : null,
+                        onPressed: widget.viewModel.deviceConnected ? widget.viewModel.disconnect : null,
                         child: const Text("Disconnect"),
                       ),
                       ElevatedButton(
-                        onPressed: widget.viewModel.deviceConnected
-                            ? discoverServices
-                            : null,
+                        onPressed: widget.viewModel.deviceConnected ? discoverServices : null,
                         child: const Text("Discover Services"),
+                      ),
+                      ElevatedButton(
+                        onPressed: widget.viewModel.deviceConnected
+                            ? readRssi
+                            : null,
+                        child: const Text("Get RSSI"),
                       ),
                     ],
                   ),
@@ -167,7 +183,7 @@ class _ServiceDiscoveryList extends StatefulWidget {
   }) : super(key: key);
 
   final String deviceId;
-  final List<DiscoveredService> discoveredServices;
+  final List<Service> discoveredServices;
 
   @override
   _ServiceDiscoveryListState createState() => _ServiceDiscoveryListState();
@@ -182,7 +198,7 @@ class _ServiceDiscoveryListState extends State<_ServiceDiscoveryList> {
     super.initState();
   }
 
-  String _charactisticsSummary(DiscoveredCharacteristic c) {
+  String _characteristicSummary(Characteristic c) {
     final props = <String>[];
     if (c.isReadable) {
       props.add("read");
@@ -203,19 +219,13 @@ class _ServiceDiscoveryListState extends State<_ServiceDiscoveryList> {
     return props.join("\n");
   }
 
-  Widget _characteristicTile(
-          DiscoveredCharacteristic characteristic, String deviceId) =>
-      ListTile(
+  Widget _characteristicTile(Characteristic characteristic) => ListTile(
         onTap: () => showDialog<void>(
-            context: context,
-            builder: (context) => CharacteristicInteractionDialog(
-                  characteristic: QualifiedCharacteristic(
-                      characteristicId: characteristic.characteristicId,
-                      serviceId: characteristic.serviceId,
-                      deviceId: deviceId),
-                )),
+          context: context,
+          builder: (context) => CharacteristicInteractionDialog(characteristic: characteristic),
+        ),
         title: Text(
-          '${characteristic.characteristicId}\n(${_charactisticsSummary(characteristic)})',
+          '${characteristic.id}\n(${_characteristicSummary(characteristic)})',
           style: const TextStyle(
             fontSize: 14,
           ),
@@ -241,19 +251,15 @@ class _ServiceDiscoveryListState extends State<_ServiceDiscoveryList> {
                       ),
                     ),
                   ),
-                  ListView.builder(
-                    shrinkWrap: true,
-                    itemBuilder: (context, index) => _characteristicTile(
-                      service.characteristics[index],
-                      widget.deviceId,
-                    ),
-                    itemCount: service.characteristicIds.length,
+                  Column(
+                    mainAxisSize: MainAxisSize.min,
+                    children: service.characteristics.map(_characteristicTile).toList(),
                   ),
                 ],
               ),
               headerBuilder: (context, isExpanded) => ListTile(
                 title: Text(
-                  '${service.serviceId}',
+                  '${service.id}',
                   style: const TextStyle(fontSize: 14),
                 ),
               ),
@@ -268,15 +274,16 @@ class _ServiceDiscoveryListState extends State<_ServiceDiscoveryList> {
   @override
   Widget build(BuildContext context) => widget.discoveredServices.isEmpty
       ? const SizedBox()
-      : Padding(
-          padding: const EdgeInsetsDirectional.only(
-            top: 20.0,
-            start: 20.0,
-            end: 20.0,
-          ),
-          child: ExpansionPanelList(
-            expansionCallback: (int index, bool isExpanded) {
-              setState(() {
+      : SafeArea(
+          top: false,
+          child: Padding(
+            padding: const EdgeInsetsDirectional.only(
+              top: 20.0,
+              start: 20.0,
+              end: 20.0,
+            ),
+            child: ExpansionPanelList(
+              expansionCallback: (int index, bool isExpanded) {
                 setState(() {
                   if (isExpanded) {
                     _expandedItems.remove(index);
@@ -284,11 +291,9 @@ class _ServiceDiscoveryListState extends State<_ServiceDiscoveryList> {
                     _expandedItems.add(index);
                   }
                 });
-              });
-            },
-            children: [
-              ...buildPanels(),
-            ],
+              },
+              children: buildPanels(),
+            ),
           ),
         );
 }
