@@ -1,14 +1,14 @@
 package com.signify.hue.reactivebleexample
 
-
 import androidx.annotation.NonNull
 import io.flutter.embedding.android.FlutterActivity
 import io.flutter.embedding.engine.FlutterEngine
 import io.flutter.plugins.GeneratedPluginRegistrant
 
-
-class MainActivity: FlutterActivity(){
-  override fun configureFlutterEngine(@NonNull flutterEngine: FlutterEngine) {
-    GeneratedPluginRegistrant.registerWith(flutterEngine)
-  }
+class MainActivity : FlutterActivity() {
+    override fun configureFlutterEngine(
+        @NonNull flutterEngine: FlutterEngine,
+    ) {
+        GeneratedPluginRegistrant.registerWith(flutterEngine)
+    }
 }
