@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.2 from annotations
+// Mocks generated by Mockito 5.4.4 from annotations
 // in flutter_reactive_ble/test/device_connector_test.dart.
 // Do not manually edit this file.
 
@@ -16,6 +16,8 @@ import 'package:reactive_ble_platform_interface/src/models.dart' as _i2;
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -80,23 +82,27 @@ class MockReactiveBlePlatform extends _i1.Mock
         Invocation.getter(#scanStream),
         returnValue: _i4.Stream<_i2.ScanResult>.empty(),
       ) as _i4.Stream<_i2.ScanResult>);
+
   @override
   _i4.Stream<_i2.BleStatus> get bleStatusStream => (super.noSuchMethod(
         Invocation.getter(#bleStatusStream),
         returnValue: _i4.Stream<_i2.BleStatus>.empty(),
       ) as _i4.Stream<_i2.BleStatus>);
+
   @override
   _i4.Stream<_i2.ConnectionStateUpdate> get connectionUpdateStream =>
       (super.noSuchMethod(
         Invocation.getter(#connectionUpdateStream),
         returnValue: _i4.Stream<_i2.ConnectionStateUpdate>.empty(),
       ) as _i4.Stream<_i2.ConnectionStateUpdate>);
+
   @override
   _i4.Stream<_i2.CharacteristicValue> get charValueUpdateStream =>
       (super.noSuchMethod(
         Invocation.getter(#charValueUpdateStream),
         returnValue: _i4.Stream<_i2.CharacteristicValue>.empty(),
       ) as _i4.Stream<_i2.CharacteristicValue>);
+
   @override
   _i4.Future<void> initialize() => (super.noSuchMethod(
         Invocation.method(
@@ -106,6 +112,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<void> deinitialize() => (super.noSuchMethod(
         Invocation.method(
@@ -115,6 +122,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<_i2.DeviceAssociationInfo?> launchCompanionWorkflow({
     required String? pattern,
@@ -133,6 +141,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Future<_i2.DeviceAssociationInfo?>.value(),
       ) as _i4.Future<_i2.DeviceAssociationInfo?>);
+
   @override
   _i4.Stream<void> scanForDevices({
     required List<_i2.Uuid>? withServices,
@@ -151,6 +160,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<_i2.Result<_i2.Unit, _i2.GenericFailure<_i2.ClearGattCacheError>?>>
       clearGattCache(String? deviceId) => (super.noSuchMethod(
@@ -170,8 +180,9 @@ class MockReactiveBlePlatform extends _i1.Mock
               ),
             )),
           ) as _i4.Future<
-              _i2.Result<_i2.Unit,
-                  _i2.GenericFailure<_i2.ClearGattCacheError>?>>);
+              _i2
+              .Result<_i2.Unit, _i2.GenericFailure<_i2.ClearGattCacheError>?>>);
+
   @override
   _i4.Future<_i2.BondingStatus> establishBonding(String? deviceId) =>
       (super.noSuchMethod(
@@ -182,6 +193,16 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue:
             _i4.Future<_i2.BondingStatus>.value(_i2.BondingStatus.none),
       ) as _i4.Future<_i2.BondingStatus>);
+
+  @override
+  _i4.Future<int> readRssi(String? deviceId) => (super.noSuchMethod(
+        Invocation.method(
+          #readRssi,
+          [deviceId],
+        ),
+        returnValue: _i4.Future<int>.value(0),
+      ) as _i4.Future<int>);
+
   @override
   _i4.Stream<void> connectToDevice(
     String? id,
@@ -199,6 +220,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<String?> retrieveDeviceName(String? id) => (super.noSuchMethod(
         Invocation.method(
@@ -207,6 +229,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Future<String?>.value(),
       ) as _i4.Future<String?>);
+
   @override
   _i4.Future<void> disconnectDevice(String? deviceId) => (super.noSuchMethod(
         Invocation.method(
@@ -216,6 +239,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<List<_i2.DiscoveredService>> discoverServices(String? deviceId) =>
       (super.noSuchMethod(
@@ -226,9 +250,22 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<List<_i2.DiscoveredService>>.value(
             <_i2.DiscoveredService>[]),
       ) as _i4.Future<List<_i2.DiscoveredService>>);
+
+  @override
+  _i4.Future<List<_i2.DiscoveredService>> getDiscoverServices(
+          String? deviceId) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getDiscoverServices,
+          [deviceId],
+        ),
+        returnValue: _i4.Future<List<_i2.DiscoveredService>>.value(
+            <_i2.DiscoveredService>[]),
+      ) as _i4.Future<List<_i2.DiscoveredService>>);
+
   @override
   _i4.Stream<void> readCharacteristic(
-          _i2.QualifiedCharacteristic? characteristic) =>
+          _i2.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #readCharacteristic,
@@ -236,9 +273,10 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<_i2.WriteCharacteristicInfo> writeCharacteristicWithResponse(
-    _i2.QualifiedCharacteristic? characteristic,
+    _i2.CharacteristicInstance? characteristic,
     List<int>? value,
   ) =>
       (super.noSuchMethod(
@@ -261,9 +299,10 @@ class MockReactiveBlePlatform extends _i1.Mock
           ),
         )),
       ) as _i4.Future<_i2.WriteCharacteristicInfo>);
+
   @override
   _i4.Future<_i2.WriteCharacteristicInfo> writeCharacteristicWithoutResponse(
-    _i2.QualifiedCharacteristic? characteristic,
+    _i2.CharacteristicInstance? characteristic,
     List<int>? value,
   ) =>
       (super.noSuchMethod(
@@ -286,9 +325,10 @@ class MockReactiveBlePlatform extends _i1.Mock
           ),
         )),
       ) as _i4.Future<_i2.WriteCharacteristicInfo>);
+
   @override
   _i4.Stream<void> subscribeToNotifications(
-          _i2.QualifiedCharacteristic? characteristic) =>
+          _i2.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #subscribeToNotifications,
@@ -296,9 +336,10 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<void> stopSubscribingToNotifications(
-          _i2.QualifiedCharacteristic? characteristic) =>
+          _i2.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #stopSubscribingToNotifications,
@@ -307,6 +348,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<int> requestMtuSize(
     String? deviceId,
@@ -322,6 +364,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Future<int>.value(0),
       ) as _i4.Future<int>);
+
   @override
   _i4.Future<_i2.ConnectionPriorityInfo> requestConnectionPriority(
     String? deviceId,
@@ -375,6 +418,7 @@ class MockDeviceScanner extends _i1.Mock implements _i5.DeviceScanner {
         ),
         returnValue: _i4.Future<_i2.DeviceAssociationInfo?>.value(),
       ) as _i4.Future<_i2.DeviceAssociationInfo?>);
+
   @override
   _i4.Stream<_i2.DiscoveredDevice> scanForDevices({
     required List<_i2.Uuid>? withServices,
@@ -412,6 +456,7 @@ class MockDiscoveredDevicesRegistry extends _i1.Mock
           Invocation.getter(#getTimestamp),
         ),
       ) as DateTime Function());
+
   @override
   void add(String? deviceId) => super.noSuchMethod(
         Invocation.method(
@@ -420,6 +465,7 @@ class MockDiscoveredDevicesRegistry extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void remove(String? deviceId) => super.noSuchMethod(
         Invocation.method(
@@ -428,6 +474,7 @@ class MockDiscoveredDevicesRegistry extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   bool isEmpty() => (super.noSuchMethod(
         Invocation.method(
@@ -436,6 +483,7 @@ class MockDiscoveredDevicesRegistry extends _i1.Mock
         ),
         returnValue: false,
       ) as bool);
+
   @override
   bool deviceIsDiscoveredRecently({
     required String? deviceId,
