@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.2 from annotations
+// Mocks generated by Mockito 5.4.4 from annotations
 // in flutter_reactive_ble/test/device_scanner_test.dart.
 // Do not manually edit this file.
 
@@ -14,6 +14,8 @@ import 'package:reactive_ble_platform_interface/src/reactive_ble_platform_interf
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -68,23 +70,27 @@ class MockReactiveBlePlatform extends _i1.Mock
         Invocation.getter(#scanStream),
         returnValue: _i4.Stream<_i2.ScanResult>.empty(),
       ) as _i4.Stream<_i2.ScanResult>);
+
   @override
   _i4.Stream<_i2.BleStatus> get bleStatusStream => (super.noSuchMethod(
         Invocation.getter(#bleStatusStream),
         returnValue: _i4.Stream<_i2.BleStatus>.empty(),
       ) as _i4.Stream<_i2.BleStatus>);
+
   @override
   _i4.Stream<_i2.ConnectionStateUpdate> get connectionUpdateStream =>
       (super.noSuchMethod(
         Invocation.getter(#connectionUpdateStream),
         returnValue: _i4.Stream<_i2.ConnectionStateUpdate>.empty(),
       ) as _i4.Stream<_i2.ConnectionStateUpdate>);
+
   @override
   _i4.Stream<_i2.CharacteristicValue> get charValueUpdateStream =>
       (super.noSuchMethod(
         Invocation.getter(#charValueUpdateStream),
         returnValue: _i4.Stream<_i2.CharacteristicValue>.empty(),
       ) as _i4.Stream<_i2.CharacteristicValue>);
+
   @override
   _i4.Future<void> initialize() => (super.noSuchMethod(
         Invocation.method(
@@ -94,6 +100,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<void> deinitialize() => (super.noSuchMethod(
         Invocation.method(
@@ -103,6 +110,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<_i2.DeviceAssociationInfo?> launchCompanionWorkflow({
     required String? pattern,
@@ -121,6 +129,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Future<_i2.DeviceAssociationInfo?>.value(),
       ) as _i4.Future<_i2.DeviceAssociationInfo?>);
+
   @override
   _i4.Stream<void> scanForDevices({
     required List<_i2.Uuid>? withServices,
@@ -139,6 +148,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<_i2.Result<_i2.Unit, _i2.GenericFailure<_i2.ClearGattCacheError>?>>
       clearGattCache(String? deviceId) => (super.noSuchMethod(
@@ -158,8 +168,9 @@ class MockReactiveBlePlatform extends _i1.Mock
               ),
             )),
           ) as _i4.Future<
-              _i2.Result<_i2.Unit,
-                  _i2.GenericFailure<_i2.ClearGattCacheError>?>>);
+              _i2
+              .Result<_i2.Unit, _i2.GenericFailure<_i2.ClearGattCacheError>?>>);
+
   @override
   _i4.Future<_i2.BondingStatus> establishBonding(String? deviceId) =>
       (super.noSuchMethod(
@@ -170,6 +181,16 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue:
             _i4.Future<_i2.BondingStatus>.value(_i2.BondingStatus.none),
       ) as _i4.Future<_i2.BondingStatus>);
+
+  @override
+  _i4.Future<int> readRssi(String? deviceId) => (super.noSuchMethod(
+        Invocation.method(
+          #readRssi,
+          [deviceId],
+        ),
+        returnValue: _i4.Future<int>.value(0),
+      ) as _i4.Future<int>);
+
   @override
   _i4.Stream<void> connectToDevice(
     String? id,
@@ -187,6 +208,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<String?> retrieveDeviceName(String? id) => (super.noSuchMethod(
         Invocation.method(
@@ -195,6 +217,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Future<String?>.value(),
       ) as _i4.Future<String?>);
+
   @override
   _i4.Future<void> disconnectDevice(String? deviceId) => (super.noSuchMethod(
         Invocation.method(
@@ -204,6 +227,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<List<_i2.DiscoveredService>> discoverServices(String? deviceId) =>
       (super.noSuchMethod(
@@ -214,9 +238,22 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<List<_i2.DiscoveredService>>.value(
             <_i2.DiscoveredService>[]),
       ) as _i4.Future<List<_i2.DiscoveredService>>);
+
+  @override
+  _i4.Future<List<_i2.DiscoveredService>> getDiscoverServices(
+          String? deviceId) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getDiscoverServices,
+          [deviceId],
+        ),
+        returnValue: _i4.Future<List<_i2.DiscoveredService>>.value(
+            <_i2.DiscoveredService>[]),
+      ) as _i4.Future<List<_i2.DiscoveredService>>);
+
   @override
   _i4.Stream<void> readCharacteristic(
-          _i2.QualifiedCharacteristic? characteristic) =>
+          _i2.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #readCharacteristic,
@@ -224,9 +261,10 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<_i2.WriteCharacteristicInfo> writeCharacteristicWithResponse(
-    _i2.QualifiedCharacteristic? characteristic,
+    _i2.CharacteristicInstance? characteristic,
     List<int>? value,
   ) =>
       (super.noSuchMethod(
@@ -249,9 +287,10 @@ class MockReactiveBlePlatform extends _i1.Mock
           ),
         )),
       ) as _i4.Future<_i2.WriteCharacteristicInfo>);
+
   @override
   _i4.Future<_i2.WriteCharacteristicInfo> writeCharacteristicWithoutResponse(
-    _i2.QualifiedCharacteristic? characteristic,
+    _i2.CharacteristicInstance? characteristic,
     List<int>? value,
   ) =>
       (super.noSuchMethod(
@@ -274,9 +313,10 @@ class MockReactiveBlePlatform extends _i1.Mock
           ),
         )),
       ) as _i4.Future<_i2.WriteCharacteristicInfo>);
+
   @override
   _i4.Stream<void> subscribeToNotifications(
-          _i2.QualifiedCharacteristic? characteristic) =>
+          _i2.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #subscribeToNotifications,
@@ -284,9 +324,10 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<void> stopSubscribingToNotifications(
-          _i2.QualifiedCharacteristic? characteristic) =>
+          _i2.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #stopSubscribingToNotifications,
@@ -295,6 +336,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<int> requestMtuSize(
     String? deviceId,
@@ -310,6 +352,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Future<int>.value(0),
       ) as _i4.Future<int>);
+
   @override
   _i4.Future<_i2.ConnectionPriorityInfo> requestConnectionPriority(
     String? deviceId,
