@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.2 from annotations
+// Mocks generated by Mockito 5.4.4 from annotations
 // in flutter_reactive_ble/test/reactive_ble_test.dart.
 // Do not manually edit this file.
 
@@ -17,6 +17,8 @@ import 'package:reactive_ble_platform_interface/src/models.dart' as _i2;
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -71,23 +73,27 @@ class MockReactiveBlePlatform extends _i1.Mock
         Invocation.getter(#scanStream),
         returnValue: _i4.Stream<_i2.ScanResult>.empty(),
       ) as _i4.Stream<_i2.ScanResult>);
+
   @override
   _i4.Stream<_i2.BleStatus> get bleStatusStream => (super.noSuchMethod(
         Invocation.getter(#bleStatusStream),
         returnValue: _i4.Stream<_i2.BleStatus>.empty(),
       ) as _i4.Stream<_i2.BleStatus>);
+
   @override
   _i4.Stream<_i2.ConnectionStateUpdate> get connectionUpdateStream =>
       (super.noSuchMethod(
         Invocation.getter(#connectionUpdateStream),
         returnValue: _i4.Stream<_i2.ConnectionStateUpdate>.empty(),
       ) as _i4.Stream<_i2.ConnectionStateUpdate>);
+
   @override
   _i4.Stream<_i2.CharacteristicValue> get charValueUpdateStream =>
       (super.noSuchMethod(
         Invocation.getter(#charValueUpdateStream),
         returnValue: _i4.Stream<_i2.CharacteristicValue>.empty(),
       ) as _i4.Stream<_i2.CharacteristicValue>);
+
   @override
   _i4.Future<void> initialize() => (super.noSuchMethod(
         Invocation.method(
@@ -97,6 +103,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<void> deinitialize() => (super.noSuchMethod(
         Invocation.method(
@@ -106,6 +113,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<_i2.DeviceAssociationInfo?> launchCompanionWorkflow({
     required String? pattern,
@@ -124,6 +132,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Future<_i2.DeviceAssociationInfo?>.value(),
       ) as _i4.Future<_i2.DeviceAssociationInfo?>);
+
   @override
   _i4.Stream<void> scanForDevices({
     required List<_i2.Uuid>? withServices,
@@ -142,6 +151,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<_i2.Result<_i2.Unit, _i2.GenericFailure<_i2.ClearGattCacheError>?>>
       clearGattCache(String? deviceId) => (super.noSuchMethod(
@@ -161,8 +171,9 @@ class MockReactiveBlePlatform extends _i1.Mock
               ),
             )),
           ) as _i4.Future<
-              _i2.Result<_i2.Unit,
-                  _i2.GenericFailure<_i2.ClearGattCacheError>?>>);
+              _i2
+              .Result<_i2.Unit, _i2.GenericFailure<_i2.ClearGattCacheError>?>>);
+
   @override
   _i4.Future<_i2.BondingStatus> establishBonding(String? deviceId) =>
       (super.noSuchMethod(
@@ -173,6 +184,16 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue:
             _i4.Future<_i2.BondingStatus>.value(_i2.BondingStatus.none),
       ) as _i4.Future<_i2.BondingStatus>);
+
+  @override
+  _i4.Future<int> readRssi(String? deviceId) => (super.noSuchMethod(
+        Invocation.method(
+          #readRssi,
+          [deviceId],
+        ),
+        returnValue: _i4.Future<int>.value(0),
+      ) as _i4.Future<int>);
+
   @override
   _i4.Stream<void> connectToDevice(
     String? id,
@@ -190,6 +211,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<String?> retrieveDeviceName(String? id) => (super.noSuchMethod(
         Invocation.method(
@@ -198,6 +220,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Future<String?>.value(),
       ) as _i4.Future<String?>);
+
   @override
   _i4.Future<void> disconnectDevice(String? deviceId) => (super.noSuchMethod(
         Invocation.method(
@@ -207,6 +230,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<List<_i2.DiscoveredService>> discoverServices(String? deviceId) =>
       (super.noSuchMethod(
@@ -217,9 +241,22 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<List<_i2.DiscoveredService>>.value(
             <_i2.DiscoveredService>[]),
       ) as _i4.Future<List<_i2.DiscoveredService>>);
+
+  @override
+  _i4.Future<List<_i2.DiscoveredService>> getDiscoverServices(
+          String? deviceId) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getDiscoverServices,
+          [deviceId],
+        ),
+        returnValue: _i4.Future<List<_i2.DiscoveredService>>.value(
+            <_i2.DiscoveredService>[]),
+      ) as _i4.Future<List<_i2.DiscoveredService>>);
+
   @override
   _i4.Stream<void> readCharacteristic(
-          _i2.QualifiedCharacteristic? characteristic) =>
+          _i2.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #readCharacteristic,
@@ -227,9 +264,10 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<_i2.WriteCharacteristicInfo> writeCharacteristicWithResponse(
-    _i2.QualifiedCharacteristic? characteristic,
+    _i2.CharacteristicInstance? characteristic,
     List<int>? value,
   ) =>
       (super.noSuchMethod(
@@ -252,9 +290,10 @@ class MockReactiveBlePlatform extends _i1.Mock
           ),
         )),
       ) as _i4.Future<_i2.WriteCharacteristicInfo>);
+
   @override
   _i4.Future<_i2.WriteCharacteristicInfo> writeCharacteristicWithoutResponse(
-    _i2.QualifiedCharacteristic? characteristic,
+    _i2.CharacteristicInstance? characteristic,
     List<int>? value,
   ) =>
       (super.noSuchMethod(
@@ -277,9 +316,10 @@ class MockReactiveBlePlatform extends _i1.Mock
           ),
         )),
       ) as _i4.Future<_i2.WriteCharacteristicInfo>);
+
   @override
   _i4.Stream<void> subscribeToNotifications(
-          _i2.QualifiedCharacteristic? characteristic) =>
+          _i2.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #subscribeToNotifications,
@@ -287,9 +327,10 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Stream<void>.empty(),
       ) as _i4.Stream<void>);
+
   @override
   _i4.Future<void> stopSubscribingToNotifications(
-          _i2.QualifiedCharacteristic? characteristic) =>
+          _i2.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #stopSubscribingToNotifications,
@@ -298,6 +339,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<int> requestMtuSize(
     String? deviceId,
@@ -313,6 +355,7 @@ class MockReactiveBlePlatform extends _i1.Mock
         ),
         returnValue: _i4.Future<int>.value(0),
       ) as _i4.Future<int>);
+
   @override
   _i4.Future<_i2.ConnectionPriorityInfo> requestConnectionPriority(
     String? deviceId,
@@ -356,11 +399,13 @@ class MockLogger extends _i1.Mock implements _i3.Logger {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.LogLevel get logLevel => (super.noSuchMethod(
         Invocation.getter(#logLevel),
         returnValue: _i2.LogLevel.none,
       ) as _i2.LogLevel);
+
   @override
   void log(Object? message) => super.noSuchMethod(
         Invocation.method(
@@ -386,9 +431,10 @@ class MockConnectedDeviceOperation extends _i1.Mock
         Invocation.getter(#characteristicValueStream),
         returnValue: _i4.Stream<_i2.CharacteristicValue>.empty(),
       ) as _i4.Stream<_i2.CharacteristicValue>);
+
   @override
   _i4.Future<List<int>> readCharacteristic(
-          _i2.QualifiedCharacteristic? characteristic) =>
+          _i2.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #readCharacteristic,
@@ -396,9 +442,10 @@ class MockConnectedDeviceOperation extends _i1.Mock
         ),
         returnValue: _i4.Future<List<int>>.value(<int>[]),
       ) as _i4.Future<List<int>>);
+
   @override
   _i4.Future<void> writeCharacteristicWithResponse(
-    _i2.QualifiedCharacteristic? characteristic, {
+    _i2.CharacteristicInstance? characteristic, {
     required List<int>? value,
   }) =>
       (super.noSuchMethod(
@@ -410,9 +457,10 @@ class MockConnectedDeviceOperation extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<void> writeCharacteristicWithoutResponse(
-    _i2.QualifiedCharacteristic? characteristic, {
+    _i2.CharacteristicInstance? characteristic, {
     required List<int>? value,
   }) =>
       (super.noSuchMethod(
@@ -424,9 +472,10 @@ class MockConnectedDeviceOperation extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Stream<List<int>> subscribeToCharacteristic(
-    _i2.QualifiedCharacteristic? characteristic,
+    _i2.CharacteristicInstance? characteristic,
     _i4.Future<void>? isDisconnected,
   ) =>
       (super.noSuchMethod(
@@ -439,6 +488,7 @@ class MockConnectedDeviceOperation extends _i1.Mock
         ),
         returnValue: _i4.Stream<List<int>>.empty(),
       ) as _i4.Stream<List<int>>);
+
   @override
   _i4.Future<int> requestMtu(
     String? deviceId,
@@ -454,6 +504,7 @@ class MockConnectedDeviceOperation extends _i1.Mock
         ),
         returnValue: _i4.Future<int>.value(0),
       ) as _i4.Future<int>);
+
   @override
   _i4.Future<List<_i2.DiscoveredService>> discoverServices(String? deviceId) =>
       (super.noSuchMethod(
@@ -464,6 +515,19 @@ class MockConnectedDeviceOperation extends _i1.Mock
         returnValue: _i4.Future<List<_i2.DiscoveredService>>.value(
             <_i2.DiscoveredService>[]),
       ) as _i4.Future<List<_i2.DiscoveredService>>);
+
+  @override
+  _i4.Future<List<_i2.DiscoveredService>> getDiscoverServices(
+          String? deviceId) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getDiscoverServices,
+          [deviceId],
+        ),
+        returnValue: _i4.Future<List<_i2.DiscoveredService>>.value(
+            <_i2.DiscoveredService>[]),
+      ) as _i4.Future<List<_i2.DiscoveredService>>);
+
   @override
   _i4.Future<void> requestConnectionPriority(
     String? deviceId,
@@ -496,6 +560,7 @@ class MockDeviceConnector extends _i1.Mock implements _i6.DeviceConnector {
         Invocation.getter(#deviceConnectionStateUpdateStream),
         returnValue: _i4.Stream<_i2.ConnectionStateUpdate>.empty(),
       ) as _i4.Stream<_i2.ConnectionStateUpdate>);
+
   @override
   _i4.Future<_i2.BondingStatus> establishBonding({required String? deviceId}) =>
       (super.noSuchMethod(
@@ -507,6 +572,7 @@ class MockDeviceConnector extends _i1.Mock implements _i6.DeviceConnector {
         returnValue:
             _i4.Future<_i2.BondingStatus>.value(_i2.BondingStatus.none),
       ) as _i4.Future<_i2.BondingStatus>);
+
   @override
   _i4.Stream<_i2.ConnectionStateUpdate> connect({
     required String? id,
@@ -526,6 +592,7 @@ class MockDeviceConnector extends _i1.Mock implements _i6.DeviceConnector {
         ),
         returnValue: _i4.Stream<_i2.ConnectionStateUpdate>.empty(),
       ) as _i4.Stream<_i2.ConnectionStateUpdate>);
+
   @override
   _i4.Future<String?> retrieveDeviceName(String? id) => (super.noSuchMethod(
         Invocation.method(
@@ -534,6 +601,7 @@ class MockDeviceConnector extends _i1.Mock implements _i6.DeviceConnector {
         ),
         returnValue: _i4.Future<String?>.value(),
       ) as _i4.Future<String?>);
+
   @override
   _i4.Stream<_i2.ConnectionStateUpdate> connectToAdvertisingDevice({
     required String? id,
@@ -585,6 +653,7 @@ class MockDeviceScanner extends _i1.Mock implements _i7.DeviceScanner {
         ),
         returnValue: _i4.Future<_i2.DeviceAssociationInfo?>.value(),
       ) as _i4.Future<_i2.DeviceAssociationInfo?>);
+
   @override
   _i4.Stream<_i2.DiscoveredDevice> scanForDevices({
     required List<_i2.Uuid>? withServices,
