@@ -25,7 +25,7 @@ class ReactiveBlePlugin : FlutterPlugin, MethodChannel.MethodCallHandler, Activi
         private fun initializePlugin(
             messenger: BinaryMessenger,
             context: Context,
-            plugin: ReactiveBlePlugin
+            plugin: ReactiveBlePlugin,
         ) {
             val channel = MethodChannel(messenger, "flutter_reactive_ble_method")
             channel.setMethodCallHandler(plugin)
@@ -39,7 +39,10 @@ class ReactiveBlePlugin : FlutterPlugin, MethodChannel.MethodCallHandler, Activi
         }
     }
 
-    override fun onMethodCall(call: MethodCall, result: Result) {
+    override fun onMethodCall(
+        call: MethodCall,
+        result: Result,
+    ) {
         pluginController.execute(call, result)
     }
 
@@ -60,6 +63,4 @@ class ReactiveBlePlugin : FlutterPlugin, MethodChannel.MethodCallHandler, Activi
     override fun onDetachedFromActivity() {
         pluginController.setActivity(null)
     }
-
-
 }
