@@ -1,6 +1,5 @@
 package com.signify.hue.flutterreactiveble.channelhandlers
 
-import com.signify.hue.flutterreactiveble.ProtobufModel as pb
 import com.signify.hue.flutterreactiveble.converters.ProtobufMessageConverter
 import com.signify.hue.flutterreactiveble.utils.Duration
 import io.flutter.plugin.common.EventChannel
@@ -8,6 +7,7 @@ import io.reactivex.Single
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
 import java.util.concurrent.TimeUnit
+import com.signify.hue.flutterreactiveble.ProtobufModel as pb
 
 class DeviceConnectionHandler(private val bleClient: com.signify.hue.flutterreactiveble.ble.BleClient) : EventChannel.StreamHandler {
     private var connectDeviceSink: EventChannel.EventSink? = null
@@ -15,7 +15,10 @@ class DeviceConnectionHandler(private val bleClient: com.signify.hue.flutterreac
 
     private lateinit var connectionUpdatesDisposable: Disposable
 
-    override fun onListen(objectSink: Any?, eventSink: EventChannel.EventSink?) {
+    override fun onListen(
+        objectSink: Any?,
+        eventSink: EventChannel.EventSink?,
+    ) {
         eventSink?.let {
             connectDeviceSink = eventSink
             connectionUpdatesDisposable = listenToConnectionChanges()
@@ -33,8 +36,8 @@ class DeviceConnectionHandler(private val bleClient: com.signify.hue.flutterreac
 
     fun connectToDevice(connectToDeviceMessage: pb.ConnectToDeviceRequest) {
         bleClient.connectToDevice(
-                connectToDeviceMessage.deviceId,
-                Duration(connectToDeviceMessage.timeoutInMs.toLong(), TimeUnit.MILLISECONDS)
+            connectToDeviceMessage.deviceId,
+            Duration(connectToDeviceMessage.timeoutInMs.toLong(), TimeUnit.MILLISECONDS),
         )
     }
 
@@ -47,7 +50,8 @@ class DeviceConnectionHandler(private val bleClient: com.signify.hue.flutterreac
         bleClient.disconnectAllDevices()
     }
 
-    private fun listenToConnectionChanges() = bleClient.connectionUpdateSubject
+    private fun listenToConnectionChanges() =
+        bleClient.connectionUpdateSubject
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe { update ->
                 when (update) {
@@ -56,7 +60,7 @@ class DeviceConnectionHandler(private val bleClient: com.signify.hue.flutterreac
                     }
                     is com.signify.hue.flutterreactiveble.ble.ConnectionUpdateError -> {
                         handleDeviceConnectionUpdateResult(
-                                converter.convertConnectionErrorToDeviceInfo(update.deviceId, update.errorMessage)
+                            converter.convertConnectionErrorToDeviceInfo(update.deviceId, update.errorMessage),
                         )
                     }
                 }
