@@ -1,7 +1,6 @@
 package com.signify.hue.flutterreactiveble.channelhandlers
 
 import android.os.ParcelUuid
-import com.signify.hue.flutterreactiveble.ProtobufModel as pb
 import com.signify.hue.flutterreactiveble.converters.ProtobufMessageConverter
 import com.signify.hue.flutterreactiveble.converters.UuidConverter
 import com.signify.hue.flutterreactiveble.model.ScanMode
@@ -9,9 +8,9 @@ import com.signify.hue.flutterreactiveble.model.createScanMode
 import io.flutter.plugin.common.EventChannel
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
+import com.signify.hue.flutterreactiveble.ProtobufModel as pb
 
 class ScanDevicesHandler(private val bleClient: com.signify.hue.flutterreactiveble.ble.BleClient) : EventChannel.StreamHandler {
-
     private var scanDevicesSink: EventChannel.EventSink? = null
     private lateinit var scanForDevicesDisposable: Disposable
     private val converter = ProtobufMessageConverter()
@@ -20,7 +19,10 @@ class ScanDevicesHandler(private val bleClient: com.signify.hue.flutterreactiveb
         private var scanParameters: ScanParameters? = null
     }
 
-    override fun onListen(objectSink: Any?, eventSink: EventChannel.EventSink?) {
+    override fun onListen(
+        objectSink: Any?,
+        eventSink: EventChannel.EventSink?,
+    ) {
         eventSink?.let {
             scanDevicesSink = eventSink
             startDeviceScan()
@@ -34,32 +36,36 @@ class ScanDevicesHandler(private val bleClient: com.signify.hue.flutterreactiveb
 
     private fun startDeviceScan() {
         scanParameters?.let { params ->
-            scanForDevicesDisposable = bleClient.scanForDevices(params.filter, params.mode, params.locationServiceIsMandatory)
+            scanForDevicesDisposable =
+                bleClient.scanForDevices(params.filter, params.mode, params.locationServiceIsMandatory)
                     .observeOn(AndroidSchedulers.mainThread())
                     .subscribe(
-                            { scanResult ->
-                                handleDeviceScanResult(converter.convertScanInfo(scanResult))
-                            },
-                            { throwable ->
-                                handleDeviceScanResult(converter.convertScanErrorInfo(throwable.message))
-                            }
+                        { scanResult ->
+                            handleDeviceScanResult(converter.convertScanInfo(scanResult))
+                        },
+                        { throwable ->
+                            handleDeviceScanResult(converter.convertScanErrorInfo(throwable.message))
+                        },
                     )
         }
-                ?: handleDeviceScanResult(converter.convertScanErrorInfo("Scanning parameters are not set"))
+            ?: handleDeviceScanResult(converter.convertScanErrorInfo("Scanning parameters are not set"))
     }
 
     fun stopDeviceScan() {
-        if (this::scanForDevicesDisposable.isInitialized) scanForDevicesDisposable.let {
-            if (!it.isDisposed) {
-                it.dispose()
-                scanParameters = null
+        if (this::scanForDevicesDisposable.isInitialized) {
+            scanForDevicesDisposable.let {
+                if (!it.isDisposed) {
+                    it.dispose()
+                    scanParameters = null
+                }
             }
         }
     }
 
     fun prepareScan(scanMessage: pb.ScanForDevicesRequest) {
         stopDeviceScan()
-        val filter = scanMessage.serviceUuidsList
+        val filter =
+            scanMessage.serviceUuidsList
                 .map { ParcelUuid(UuidConverter().uuidFromByteArray(it.data.toByteArray())) }
         val scanMode = createScanMode(scanMessage.scanMode)
         scanParameters = ScanParameters(filter, scanMode, scanMessage.requireLocationServicesEnabled)
