@@ -11,7 +11,6 @@ class UuidConverter {
     }
 
     fun uuidFromByteArray(bytes: ByteArray): UUID {
-
         return when (bytes.size) {
             byteSize16Bit -> convert16BitToUuid(bytes)
             byteSize32Bit -> convert32BitToUuid(bytes)
@@ -22,16 +21,50 @@ class UuidConverter {
     @Suppress("Detekt.MagicNumber")
     private fun convert16BitToUuid(bytes: ByteArray): UUID {
         // UUID construction is retrieved from BLE corespec v5.0 page 1917
-        val uuidConstruct = byteArrayOf(0x00, 0x00, bytes[0], bytes[1], 0x00, 0x00, 0x10, 0x00,
-                0x80.toByte(), 0x00, 0x00, 0x80.toByte(), 0x5F, 0x9B.toByte(), 0x34, 0xFB.toByte())
+        val uuidConstruct =
+            byteArrayOf(
+                0x00,
+                0x00,
+                bytes[0],
+                bytes[1],
+                0x00,
+                0x00,
+                0x10,
+                0x00,
+                0x80.toByte(),
+                0x00,
+                0x00,
+                0x80.toByte(),
+                0x5F,
+                0x9B.toByte(),
+                0x34,
+                0xFB.toByte(),
+            )
 
         return convert128BitNotationToUuid(uuidConstruct)
     }
 
     @Suppress("Detekt.MagicNumber")
     private fun convert32BitToUuid(bytes: ByteArray): UUID {
-        val uuidConstruct = byteArrayOf(bytes[0], bytes[1], bytes[2], bytes[3], 0x00, 0x00, 0x10, 0x00,
-                0x80.toByte(), 0x00, 0x00, 0x80.toByte(), 0x5F, 0x9B.toByte(), 0x34, 0xFB.toByte())
+        val uuidConstruct =
+            byteArrayOf(
+                bytes[0],
+                bytes[1],
+                bytes[2],
+                bytes[3],
+                0x00,
+                0x00,
+                0x10,
+                0x00,
+                0x80.toByte(),
+                0x00,
+                0x00,
+                0x80.toByte(),
+                0x5F,
+                0x9B.toByte(),
+                0x34,
+                0xFB.toByte(),
+            )
 
         return convert128BitNotationToUuid(uuidConstruct)
     }
