@@ -8,7 +8,7 @@ enum class ConnectionState(val code: Int) {
     CONNECTED(1),
     DISCONNECTING(2),
     DISCONNECTED(3),
-    UNKNOWN(4)
+    UNKNOWN(4),
 }
 
 fun RxBleConnection.RxBleConnectionState.toConnectionState(): ConnectionState =
@@ -18,4 +18,4 @@ fun RxBleConnection.RxBleConnectionState.toConnectionState(): ConnectionState =
         "CONNECTED" -> ConnectionState.CONNECTED
         "DISCONNECTING" -> ConnectionState.DISCONNECTING
         else -> ConnectionState.UNKNOWN
-}
+    }
