@@ -2,12 +2,14 @@ package com.signify.hue.flutterreactiveble.ble
 
 import android.bluetooth.BluetoothDevice
 import android.bluetooth.BluetoothDevice.BOND_BONDED
+import android.bluetooth.BluetoothGattCharacteristic
 import android.content.Context
 import com.google.common.truth.Truth.assertThat
 import com.polidea.rxandroidble2.RxBleClient
 import com.polidea.rxandroidble2.RxBleConnection
 import com.polidea.rxandroidble2.RxBleDevice
 import com.polidea.rxandroidble2.RxBleDeviceServices
+import com.signify.hue.flutterreactiveble.ble.extensions.resolveCharacteristic
 import com.signify.hue.flutterreactiveble.ble.extensions.writeCharWithResponse
 import com.signify.hue.flutterreactiveble.ble.extensions.writeCharWithoutResponse
 import com.signify.hue.flutterreactiveble.utils.Duration
@@ -32,19 +34,22 @@ import org.junit.jupiter.api.Test
 import java.util.UUID
 import java.util.concurrent.TimeUnit
 
-private class BleClientForTesting(val bleClient: RxBleClient, appContext: Context, val deviceConnector: DeviceConnector) : ReactiveBleClient(appContext) {
-
+private class BleClientForTesting(val bleClient: RxBleClient, appContext: Context, val deviceConnector: DeviceConnector) : ReactiveBleClient(
+    appContext,
+) {
     override fun initializeClient() {
         rxBleClient = bleClient
         activeConnections = mutableMapOf()
     }
 
-    override fun createDeviceConnector(device: RxBleDevice, timeout: Duration): DeviceConnector = deviceConnector
+    override fun createDeviceConnector(
+        device: RxBleDevice,
+        timeout: Duration,
+    ): DeviceConnector = deviceConnector
 }
 
 @DisplayName("BleClient unit tests")
 class ReactiveBleClientTest {
-
     @MockK
     private lateinit var context: Context
 
@@ -104,11 +109,9 @@ class ReactiveBleClientTest {
         subject.onComplete()
     }
 
-
     @DisplayName("Establishing a connection")
     @Nested
     inner class EstablishConnectionTest {
-
         @Test
         fun `should use deviceconnector when connecting to a device`() {
             sut.connectToDevice("test", testTimeout)
@@ -122,36 +125,55 @@ class ReactiveBleClientTest {
     inner class BleOperationsTest {
         @Test
         fun `should call readcharacteristic in case the connection is established`() {
-            sut.readCharacteristic("test", UUID.randomUUID()).test()
+            every {
+                rxConnection.resolveCharacteristic(
+                    any(),
+                    any(),
+                )
+            }.returns(Single.just(BluetoothGattCharacteristic(UUID.randomUUID(), 0, 0)))
 
-            verify(exactly = 1) { rxConnection.readCharacteristic(any<UUID>()) }
+            sut.readCharacteristic("test", UUID.randomUUID(), 11).test()
+
+            verify(exactly = 1) { rxConnection.readCharacteristic(any<BluetoothGattCharacteristic>()) }
         }
 
         @Test
         fun `should not call readcharacteristic in case the connection is not established`() {
             subject.onNext(EstablishConnectionFailure("test", "error"))
 
-            sut.readCharacteristic("test", UUID.randomUUID()).test()
+            val result = sut.readCharacteristic("test", UUID.randomUUID(), 11).test()
 
-            verify(exactly = 0) { rxConnection.readCharacteristic(any<UUID>()) }
+            assertThat(result.values().first()).isInstanceOf(CharOperationFailed::class.java)
+            verify(exactly = 0) { rxConnection.readCharacteristic(any<BluetoothGattCharacteristic>()) }
         }
 
         @Test
         fun `should report failure in case reading characteristic fails`() {
             subject.onNext(EstablishConnectionFailure("test", "error"))
 
-            val observable = sut.readCharacteristic("test", UUID.randomUUID()).test()
+            val observable = sut.readCharacteristic("test", UUID.randomUUID(), 11).test()
 
             assertThat(observable.values().first()).isInstanceOf(CharOperationFailed::class.java)
         }
 
         @Test
-        fun `should incorporate the value in case readcharacteristics succeeds`() {
+        fun `should incorporate the value in case readcharacteristic succeeds`() {
             val byteMin = Byte.MIN_VALUE
             val byteMax = Byte.MAX_VALUE
 
-            every { rxConnection.readCharacteristic(any<UUID>()) }.returns(Single.just(byteArrayOf(byteMin, byteMax)))
-            val observable = sut.readCharacteristic("test", UUID.randomUUID())
+            every {
+                rxConnection.readCharacteristic(
+                    any<BluetoothGattCharacteristic>(),
+                )
+            }.returns(Single.just(byteArrayOf(byteMin, byteMax)))
+            every {
+                rxConnection.resolveCharacteristic(
+                    any(),
+                    any(),
+                )
+            }.returns(Single.just(BluetoothGattCharacteristic(UUID.randomUUID(), 0, 0)))
+            val observable =
+                sut.readCharacteristic("test", UUID.randomUUID(), 11)
                     .map { result -> result as CharOperationSuccessful }.test()
 
             assertThat(observable.values().first().value).isEqualTo(listOf(byteMin, byteMax))
@@ -163,9 +185,15 @@ class ReactiveBleClientTest {
             val byteMax = Byte.MAX_VALUE
             val bytes = byteArrayOf(byteMin, byteMax)
 
-            sut.writeCharacteristicWithResponse("test", UUID.randomUUID(), bytes).test()
+            every {
+                rxConnection.resolveCharacteristic(
+                    any(),
+                    any(),
+                )
+            }.returns(Single.just(BluetoothGattCharacteristic(UUID.randomUUID(), 0, 0)))
+            sut.writeCharacteristicWithResponse("test", UUID.randomUUID(), 11, bytes).test()
 
-            verify(exactly = 1) { rxConnection.writeCharWithResponse(any<UUID>(), any()) }
+            verify(exactly = 1) { rxConnection.writeCharWithResponse(any(), any()) }
         }
 
         @Test
@@ -174,9 +202,9 @@ class ReactiveBleClientTest {
             val byteMax = Byte.MAX_VALUE
             val bytes = byteArrayOf(byteMin, byteMax)
 
-            sut.writeCharacteristicWithResponse("test", UUID.randomUUID(), bytes).test()
+            sut.writeCharacteristicWithResponse("test", UUID.randomUUID(), 11, bytes).test()
 
-            verify(exactly = 0) { rxConnection.writeCharWithoutResponse(any<UUID>(), any()) }
+            verify(exactly = 0) { rxConnection.writeCharWithoutResponse(any(), any()) }
         }
 
         @Test
@@ -186,9 +214,10 @@ class ReactiveBleClientTest {
             val bytes = byteArrayOf(byteMin, byteMax)
             subject.onNext(EstablishConnectionFailure("test", "error"))
 
-            sut.writeCharacteristicWithResponse("test", UUID.randomUUID(), bytes).test()
+            val result = sut.writeCharacteristicWithResponse("test", UUID.randomUUID(), 11, bytes).test()
+            assertThat(result.values().first()).isInstanceOf(CharOperationFailed::class.java)
 
-            verify(exactly = 0) { rxConnection.writeCharWithResponse(any<UUID>(), any()) }
+            verify(exactly = 0) { rxConnection.writeCharWithResponse(any(), any()) }
         }
 
         @Test
@@ -197,9 +226,15 @@ class ReactiveBleClientTest {
             val byteMax = Byte.MAX_VALUE
             val bytes = byteArrayOf(byteMin, byteMax)
 
-            sut.writeCharacteristicWithoutResponse("test", UUID.randomUUID(), bytes).test()
+            every {
+                rxConnection.resolveCharacteristic(
+                    any(),
+                    any(),
+                )
+            }.returns(Single.just(BluetoothGattCharacteristic(UUID.randomUUID(), 0, 0)))
+            sut.writeCharacteristicWithoutResponse("test", UUID.randomUUID(), 11, bytes).test()
 
-            verify(exactly = 1) { rxConnection.writeCharWithoutResponse(any<UUID>(), any()) }
+            verify(exactly = 1) { rxConnection.writeCharWithoutResponse(any(), any()) }
         }
 
         @Test
@@ -209,10 +244,10 @@ class ReactiveBleClientTest {
             val bytes = byteArrayOf(byteMin, byteMax)
             subject.onNext(EstablishConnectionFailure("test", "error"))
 
+            val result = sut.writeCharacteristicWithoutResponse("test", UUID.randomUUID(), 11, bytes).test()
 
-            sut.writeCharacteristicWithoutResponse("test", UUID.randomUUID(), bytes).test()
-
-            verify(exactly = 0) { rxConnection.writeCharWithoutResponse(any<UUID>(), any()) }
+            assertThat(result.values().first()).isInstanceOf(CharOperationFailed::class.java)
+            verify(exactly = 0) { rxConnection.writeCharWithoutResponse(any(), any()) }
         }
 
         @Test
@@ -223,7 +258,7 @@ class ReactiveBleClientTest {
 
             subject.onNext(EstablishConnectionFailure("test", "error"))
 
-            val observable = sut.writeCharacteristicWithResponse("test", UUID.randomUUID(), bytes).test()
+            val observable = sut.writeCharacteristicWithResponse("test", UUID.randomUUID(), 11, bytes).test()
 
             assertThat(observable.values().first()).isInstanceOf(CharOperationFailed::class.java)
         }
@@ -235,7 +270,14 @@ class ReactiveBleClientTest {
             val bytes = byteArrayOf(byteMin, byteMax)
 
             every { rxConnection.writeCharWithResponse(any(), any()) }.returns(Single.just(byteArrayOf(byteMin, byteMax)))
-            val observable = sut.writeCharacteristicWithResponse("test", UUID.randomUUID(), bytes)
+            every {
+                rxConnection.resolveCharacteristic(
+                    any(),
+                    any(),
+                )
+            }.returns(Single.just(BluetoothGattCharacteristic(UUID.randomUUID(), 0, 0)))
+            val observable =
+                sut.writeCharacteristicWithResponse("test", UUID.randomUUID(), 11, bytes)
                     .map { result -> result as CharOperationSuccessful }.test()
 
             assertThat(observable.values().first().value).isEqualTo(bytes.toList())
@@ -245,7 +287,6 @@ class ReactiveBleClientTest {
     @Nested
     @DisplayName("Negotiate mtu")
     inner class NegotiateMtuTest {
-
         @Test
         fun `should return mtunegotiatesuccesful in case it succeeds`() {
             val mtuSize = 19
@@ -253,7 +294,7 @@ class ReactiveBleClientTest {
 
             val result = sut.negotiateMtuSize("", mtuSize).test()
 
-            assertThat(result.values().first()).isInstanceOf(MtuNegotiateSuccesful::class.java)
+            assertThat(result.values().first()).isInstanceOf(MtuNegotiateSuccessful::class.java)
         }
 
         @Test
@@ -264,6 +305,29 @@ class ReactiveBleClientTest {
             val result = sut.negotiateMtuSize("", mtuSize).test()
 
             assertThat(result.values().first()).isInstanceOf(MtuNegotiateFailed::class.java)
+        }
+    }
+
+    @Nested
+    @DisplayName("Read RSSI")
+    inner class ReadRssiTest {
+        @Test
+        fun `should return RSSI in case it succeeds`() {
+            val rssi = -42
+            every { rxConnection.readRssi() }.returns(Single.just(rssi))
+
+            val result = sut.readRssi("").test()
+
+            assertThat(result.values().first()).isEqualTo(rssi)
+        }
+
+        @Test
+        fun `should return error in case it fails`() {
+            every { rxConnection.readRssi() }.returns(Single.error(IllegalStateException("boom")))
+
+            val result = sut.readRssi("").test()
+
+            assertThat(result.errors().first()).isInstanceOf(IllegalStateException::class.java)
         }
     }
 
@@ -289,7 +353,6 @@ class ReactiveBleClientTest {
 
         @Test
         fun `starts with current state`() {
-
             val result = sut.observeBleStatus().test()
             assertThat(result.values().count()).isEqualTo(2)
             assertThat(result.values().first()).isEqualTo(BleStatus.POWERED_OFF)
@@ -299,9 +362,8 @@ class ReactiveBleClientTest {
     @Nested
     @DisplayName("Change priority")
     inner class ChangePriorityTest {
-
         @Test
-        fun `returns prioritysuccess when  completed`() {
+        fun `returns prioritysuccess when completed`() {
             val completer = Completable.fromCallable { true }
 
             every { rxConnection.requestConnectionPriority(any(), any(), any()) }.returns(completer)
@@ -320,11 +382,10 @@ class ReactiveBleClientTest {
     @Nested
     @DisplayName("Discover services")
     inner class DiscoverServicesTest {
-
         @BeforeEach
         fun setup() {
             every { bleDevice.bluetoothDevice }.returns(bluetoothDevice)
-            every {bluetoothDevice.bondState}.returns(BOND_BONDED)
+            every { bluetoothDevice.bondState }.returns(BOND_BONDED)
         }
 
         @Test
