@@ -7,7 +7,7 @@ import com.signify.hue.flutterreactiveble.ProtobufModel.EstablishBondingInfo.Bon
 import com.signify.hue.flutterreactiveble.ble.Connectable
 import com.signify.hue.flutterreactiveble.ble.ConnectionUpdateSuccess
 import com.signify.hue.flutterreactiveble.ble.MtuNegotiateFailed
-import com.signify.hue.flutterreactiveble.ble.MtuNegotiateSuccesful
+import com.signify.hue.flutterreactiveble.ble.MtuNegotiateSuccessful
 import com.signify.hue.flutterreactiveble.ble.ScanInfo
 import com.signify.hue.flutterreactiveble.model.NegotiateMtuErrorType
 import org.junit.jupiter.api.DisplayName
@@ -19,14 +19,12 @@ import java.util.UUID
 import java.util.stream.Stream
 import com.signify.hue.flutterreactiveble.ProtobufModel as pb
 
-
 class ProtobufMessageConverterTest {
     val protobufConverter = ProtobufMessageConverter()
 
     @Nested
     @DisplayName("Convert to scaninfo")
     inner class ScanInfoTest {
-
         @Test
         fun `converts scan result to DeviceDiscoveryMessage`() {
             val scanInfo = createScanInfo()
@@ -113,7 +111,6 @@ class ProtobufMessageConverterTest {
     @Nested
     @DisplayName("Convert to deviceinfo")
     inner class DeviceInfoTest {
-
         @Test
         fun `converts device id as parameter in device connection message`() {
             val deviceId = "2"
@@ -126,7 +123,7 @@ class ProtobufMessageConverterTest {
             val result = 0
             val connection = ConnectionUpdateSuccess("", result)
             assertThat(protobufConverter.convertToDeviceInfo(connection).connectionState).isEqualTo(
-                result
+                result,
             )
         }
     }
@@ -134,23 +131,17 @@ class ProtobufMessageConverterTest {
     @Nested
     @DisplayName("Convert to charinfo")
     inner class ConvertCharInfoTest {
-
         @Test
         fun `converts to a characteristicvalueInfo object `() {
             val request = createCharacteristicRequest("a", UUID.randomUUID())
 
-            assertThat(
-                protobufConverter.convertCharacteristicInfo(
-                    request.characteristic,
-                    byteArrayOf(1)
-                )
-            )
+            assertThat(protobufConverter.convertCharacteristicInfo(request.characteristic, byteArrayOf(1)))
                 .isInstanceOf(pb.CharacteristicValueInfo::class.java)
         }
 
         @Test
         fun `converts a char value and request into a characteristic info value `() {
-            val request = createCharacteristicRequest("a", UUID.randomUUID())
+            val request = createCharacteristicRequest("b", UUID.randomUUID())
             val expectedValue = byteArrayOf(1)
             val valueInfo =
                 protobufConverter.convertCharacteristicInfo(request.characteristic, expectedValue)
@@ -162,32 +153,29 @@ class ProtobufMessageConverterTest {
     @Nested
     @DisplayName("Convert to negotiatemtuinfo")
     inner class NegotiateMtuInfoTest {
-
         @Test
         fun `converts to negotiatemtuinfo object`() {
-            val result = MtuNegotiateSuccesful("", 3)
+            val result = MtuNegotiateSuccessful("", 3)
 
             assertThat(protobufConverter.convertNegotiateMtuInfo(result)).isInstanceOf(pb.NegotiateMtuInfo::class.java)
         }
 
         @Test
         fun `converts deviceId`() {
-            val result = MtuNegotiateSuccesful("id", 3)
+            val result = MtuNegotiateSuccessful("id", 3)
             assertThat(protobufConverter.convertNegotiateMtuInfo(result).deviceId).isEqualTo(result.deviceId)
         }
 
         @Test
         fun `converts mtusize`() {
-            val result = MtuNegotiateSuccesful("id", 3)
+            val result = MtuNegotiateSuccessful("id", 3)
             assertThat(protobufConverter.convertNegotiateMtuInfo(result).mtuSize).isEqualTo(result.size)
         }
 
         @Test
         fun `sets default value for error in case no error occurred`() {
-            val result = MtuNegotiateSuccesful("id", 3)
-            assertThat(protobufConverter.convertNegotiateMtuInfo(result).failure.message).isEqualTo(
-                ""
-            )
+            val result = MtuNegotiateSuccessful("id", 3)
+            assertThat(protobufConverter.convertNegotiateMtuInfo(result).failure.message).isEqualTo("")
         }
 
         @Test
@@ -210,7 +198,10 @@ class ProtobufMessageConverterTest {
     inner class BondInfoTest {
         @Test
         @MethodSource("provideParameters")
-        fun `converts bonded`(bondState: BondState, androidConstant: Int) {
+        fun `converts bonded`(
+            bondState: BondState,
+            androidConstant: Int,
+        ) {
             assertThat(protobufConverter.convertBondingInfo(androidConstant)).isEqualTo(bondState)
         }
 
@@ -218,7 +209,7 @@ class ProtobufMessageConverterTest {
             return Stream.of(
                 Arguments.of(BondState.BONDED, BluetoothDevice.BOND_BONDED),
                 Arguments.of(BondState.BONDING, BluetoothDevice.BOND_BONDING),
-                Arguments.of(BondState.NONE, BluetoothDevice.BOND_NONE)
+                Arguments.of(BondState.NONE, BluetoothDevice.BOND_NONE),
             )
         }
 
@@ -231,7 +222,7 @@ class ProtobufMessageConverterTest {
     }
 
     private fun createScanInfo(): ScanInfo {
-        val macAdress = "123"
+        val macAddress = "123"
         val deviceName = "Testdevice"
         val rssi = 200
         val uuid = UUID.randomUUID()
@@ -244,7 +235,7 @@ class ProtobufMessageConverterTest {
         val manufacturerData = "123".toByteArray()
 
         return ScanInfo(
-            deviceId = macAdress,
+            deviceId = macAddress,
             name = deviceName,
             rssi = rssi,
             connectable = Connectable.UNKNOWN,
@@ -256,20 +247,21 @@ class ProtobufMessageConverterTest {
 
     private fun createCharacteristicRequest(
         deviceId: String,
-        serviceUuid: UUID
+        serviceUuid: UUID,
     ): pb.ReadCharacteristicRequest {
         val uuidConverter = UuidConverter()
-        val uuid = pb.Uuid.newBuilder()
-            .setData(ByteString.copyFrom(uuidConverter.byteArrayFromUuid(serviceUuid)))
+        val uuid =
+            pb.Uuid.newBuilder()
+                .setData(ByteString.copyFrom(uuidConverter.byteArrayFromUuid(serviceUuid)))
 
-        val characteristicAddress = pb.CharacteristicAddress.newBuilder()
-            .setDeviceId(deviceId)
-            .setServiceUuid(uuid)
-            .setCharacteristicUuid(uuid)
+        val characteristicAddress =
+            pb.CharacteristicAddress.newBuilder()
+                .setDeviceId(deviceId)
+                .setServiceUuid(uuid)
+                .setCharacteristicUuid(uuid)
 
         return pb.ReadCharacteristicRequest.newBuilder()
             .setCharacteristic(characteristicAddress)
             .build()
     }
 }
-
