@@ -2,11 +2,9 @@ package com.signify.hue.flutterreactiveble.converters
 
 import com.google.common.truth.Truth.assertThat
 import org.junit.jupiter.api.Test
-import java.util.*
-
+import java.util.UUID
 
 class UuidConverterTest {
-
     val converter = UuidConverter()
 
     @Test
@@ -20,7 +18,6 @@ class UuidConverterTest {
 
     @Test
     fun `should be able to convert 16bit uuid`() {
-
         val array = byteArrayOf(0xFE.toByte(), 0x0F.toByte())
         val uuid = converter.uuidFromByteArray(array)
         assertThat(uuid.toString().toUpperCase()).isEqualTo("0000FE0F-0000-1000-8000-00805F9B34FB")
@@ -28,7 +25,6 @@ class UuidConverterTest {
 
     @Test
     fun `should be able to convert 32bit uuid`() {
-
         val array = byteArrayOf(0xFE.toByte(), 0x0F.toByte(), 0x0F.toByte(), 0xFE.toByte())
         val uuid = converter.uuidFromByteArray(array)
         assertThat(uuid.toString().toUpperCase()).isEqualTo("FE0F0FFE-0000-1000-8000-00805F9B34FB")
