@@ -114,7 +114,9 @@ final class PluginController {
                 let message = CharacteristicValueInfo.with {
                     $0.characteristic = CharacteristicAddress.with {
                         $0.characteristicUuid = Uuid.with { $0.data = characteristic.id.data }
+                        $0.characteristicInstanceID = characteristic.instanceID
                         $0.serviceUuid = Uuid.with { $0.data = characteristic.serviceID.data }
+                        $0.serviceInstanceID = characteristic.serviceInstanceID
                         $0.deviceID = characteristic.peripheralID.uuidString
                     }
                     if let value = value {
@@ -134,7 +136,6 @@ final class PluginController {
                     // In case message arrives before sink is created
                     context.messageQueue.append(message)
                 }
-
             }
         )
 
@@ -316,12 +317,14 @@ final class PluginController {
         func makeDiscoveredService(service: CBService) -> DiscoveredService {
             DiscoveredService.with {
                 $0.serviceUuid = Uuid.with { $0.data = service.uuid.data }
+                $0.serviceInstanceID = service.instanceId?.description ?? ""
                 $0.characteristicUuids = (service.characteristics ?? []).map { characteristic in
                     Uuid.with { $0.data = characteristic.uuid.data }
                 }
                 $0.characteristics = (service.characteristics ?? []).map { characteristic in
                     DiscoveredCharacteristic.with {
                         $0.characteristicID = Uuid.with {$0.data = characteristic.uuid.data}
+                        $0.characteristicInstanceID = characteristic.instanceId?.description ?? ""
                         if characteristic.service?.uuid.data != nil {
                             $0.serviceID = Uuid.with {$0.data = characteristic.service!.uuid.data}
                         }
@@ -353,6 +356,57 @@ final class PluginController {
         }
     }
 
+    func getDiscoveredServices(name: String, args: DiscoverServicesRequest, completion: @escaping PlatformMethodCompletionHandler) {
+        guard let central = central
+        else {
+            completion(.failure(PluginError.notInitialized.asFlutterError))
+            return
+        }
+
+        guard let deviceID = UUID(uuidString: args.deviceID)
+        else {
+            completion(.failure(PluginError.invalidMethodCall(method: name, details: "\"deviceID\" is invalid").asFlutterError))
+            return
+        }
+
+        func makeDiscoveredService(service: CBService) -> DiscoveredService {
+            DiscoveredService.with {
+                $0.serviceUuid = Uuid.with { $0.data = service.uuid.data }
+                $0.serviceInstanceID = service.instanceId?.description ?? ""
+                $0.characteristicUuids = (service.characteristics ?? []).map { characteristic in
+                    Uuid.with { $0.data = characteristic.uuid.data }
+                }
+                $0.characteristics = (service.characteristics ?? []).map { characteristic in
+                    DiscoveredCharacteristic.with {
+                        $0.characteristicID = Uuid.with {$0.data = characteristic.uuid.data}
+                        $0.characteristicInstanceID = characteristic.instanceId?.description ?? ""
+                        if characteristic.service?.uuid.data != nil {
+                            $0.serviceID = Uuid.with {$0.data = characteristic.service!.uuid.data}
+                        }
+                        $0.isReadable = characteristic.properties.contains(.read)
+                        $0.isWritableWithResponse = characteristic.properties.contains(.write)
+                        $0.isWritableWithoutResponse = characteristic.properties.contains(.writeWithoutResponse)
+                        $0.isNotifiable = characteristic.properties.contains(.notify)
+                        $0.isIndicatable = characteristic.properties.contains(.indicate)
+                    }
+                }
+
+                $0.includedServices = (service.includedServices ?? []).map(makeDiscoveredService)
+            }
+        }
+
+        do {
+            let peripheral = try central.peripheral(for: deviceID)
+            completion(.success(DiscoverServicesInfo.with {
+                        $0.deviceID = deviceID.uuidString
+                        $0.services = (peripheral.services ?? []).map(makeDiscoveredService)
+                    }))
+
+        } catch {
+            completion(.failure(PluginError.unknown(error).asFlutterError))
+        }
+    }
+
     func enableCharacteristicNotifications(name: String, args: NotifyCharacteristicRequest, completion: @escaping PlatformMethodCompletionHandler) {
         guard let central = central
         else {
@@ -360,7 +414,7 @@ final class PluginController {
             return
         }
 
-        guard let characteristic = QualifiedCharacteristicIDFactory().make(from: args.characteristic)
+        guard let characteristic = CharacteristicInstanceIDFactory().make(from: args.characteristic)
         else {
             completion(.failure(PluginError.invalidMethodCall(method: name, details: "characteristic, service, and peripheral IDs are required").asFlutterError))
             return
@@ -386,7 +440,7 @@ final class PluginController {
             return
         }
 
-        guard let characteristic = QualifiedCharacteristicIDFactory().make(from: args.characteristic)
+        guard let characteristic = CharacteristicInstanceIDFactory().make(from: args.characteristic)
         else {
             completion(.failure(PluginError.invalidMethodCall(method: name, details: "characteristic, service, and peripheral IDs are required").asFlutterError))
             return
@@ -412,7 +466,7 @@ final class PluginController {
             return
         }
 
-        guard let characteristic = QualifiedCharacteristicIDFactory().make(from: args.characteristic)
+        guard let characteristic = CharacteristicInstanceIDFactory().make(from: args.characteristic)
         else {
             completion(.failure(PluginError.invalidMethodCall(method: name, details: "characteristic, service, and peripheral IDs are required").asFlutterError))
             return
@@ -447,7 +501,7 @@ final class PluginController {
             return
         }
 
-        guard let characteristic = QualifiedCharacteristicIDFactory().make(from: args.characteristic)
+        guard let characteristic = CharacteristicInstanceIDFactory().make(from: args.characteristic)
         else {
             completion(.failure(PluginError.invalidMethodCall(method: name, details: "characteristic, service, and peripheral IDs are required").asFlutterError))
             return
@@ -456,14 +510,10 @@ final class PluginController {
         do {
             try central.writeWithResponse(
                 value: args.value,
-                characteristic: QualifiedCharacteristic(id: characteristic.id, serviceID: characteristic.serviceID, peripheralID: characteristic.peripheralID),
+                characteristic: characteristic,
                 completion: { _, characteristic, error in
                     let result = WriteCharacteristicInfo.with {
-                        $0.characteristic = CharacteristicAddress.with {
-                            $0.characteristicUuid = Uuid.with { $0.data = characteristic.id.data }
-                            $0.serviceUuid = Uuid.with { $0.data = characteristic.serviceID.data }
-                            $0.deviceID = characteristic.peripheralID.uuidString
-                        }
+                        $0.characteristic = args.characteristic
                         if let error = error {
                             $0.failure = GenericFailure.with {
                                 $0.code = Int32(WriteCharacteristicFailure.unknown.rawValue)
@@ -495,7 +545,7 @@ final class PluginController {
             return
         }
 
-        guard let characteristic = QualifiedCharacteristicIDFactory().make(from: args.characteristic)
+        guard let characteristic = CharacteristicInstanceIDFactory().make(from: args.characteristic)
         else {
             completion(.failure(PluginError.invalidMethodCall(method: name, details: "characteristic, service, and peripheral IDs are required").asFlutterError))
             return
@@ -505,7 +555,7 @@ final class PluginController {
         do {
             try central.writeWithoutResponse(
                 value: args.value,
-                characteristic: QualifiedCharacteristic(id: characteristic.id, serviceID: characteristic.serviceID, peripheralID: characteristic.peripheralID)
+                characteristic: characteristic
             )
             result = WriteCharacteristicInfo.with {
                 $0.characteristic = args.characteristic
@@ -554,6 +604,50 @@ final class PluginController {
         }
 
         completion(.success(result))
+    }
+    
+    func readRssi(name: String, args: ReadRssiRequest, completion: @escaping PlatformMethodCompletionHandler) {
+        guard let central = central
+        else {
+            completion(.failure(PluginError.notInitialized.asFlutterError))
+            return
+        }
+        
+        guard let peripheralID = UUID(uuidString: args.deviceID)
+        else {
+            completion(.failure(PluginError.invalidMethodCall(method: name, details: "peripheral ID is required").asFlutterError))
+            return
+        }
+
+        do {
+            try central.readRssi(
+                for: peripheralID,
+                completion: papply(weak: self) { context, result in
+                    result.iif(
+                        success: { rssi in
+                            let result: ReadRssiResult = ReadRssiResult.with {
+                                $0.rssi = Int32(rssi)
+                            }
+                            completion(.success(result))
+                        },
+                        failure: { error in
+                            completion(.failure(context.makeFlutterError(error: error)))
+                        }
+                    )
+                }
+            )
+        } catch let error {
+            completion(.failure(makeFlutterError(error: error)))
+        }
+    }
+
+    // takes an error and converts it into a Flutter error
+    private func makeFlutterError(error: Error) -> FlutterError {
+        if let error = error as? PluginError {
+            return error.asFlutterError
+        } else {
+            return PluginError.unknown(error).asFlutterError
+        }
     }
 
     private func reportState(_ knownState: CBManagerState? = nil) {
