@@ -16,20 +16,20 @@ abstract class ArgsToProtobufConverter {
   pb.DisconnectFromDeviceRequest createDisconnectDeviceArgs(String deviceId);
 
   pb.ReadCharacteristicRequest createReadCharacteristicRequest(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
   );
 
   pb.WriteCharacteristicRequest createWriteCharacteristicRequest(
-    QualifiedCharacteristic characteristic,
+      CharacteristicInstance characteristic,
     List<int> value,
   );
 
   pb.NotifyCharacteristicRequest createNotifyCharacteristicRequest(
-    QualifiedCharacteristic characteristic,
+      CharacteristicInstance characteristic,
   );
 
   pb.NotifyNoMoreCharacteristicRequest createNotifyNoMoreCharacteristicRequest(
-    QualifiedCharacteristic characteristic,
+      CharacteristicInstance characteristic,
   );
 
   pb.NegotiateMtuRequest createNegotiateMtuRequest(
@@ -57,6 +57,8 @@ abstract class ArgsToProtobufConverter {
   pb.ClearGattCacheRequest createClearGattCacheRequest(String deviceId);
 
   pb.DiscoverServicesRequest createDiscoverServicesRequest(String deviceId);
+
+  pb.ReadRssiRequest createReadRssiRequest(String deviceId);
 }
 
 class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
@@ -107,29 +109,33 @@ class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
 
   @override
   pb.ReadCharacteristicRequest createReadCharacteristicRequest(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
   ) {
     final args = pb.ReadCharacteristicRequest()
       ..characteristic = (pb.CharacteristicAddress()
         ..deviceId = characteristic.deviceId
         ..serviceUuid = (pb.Uuid()..data = characteristic.serviceId.data)
+        ..serviceInstanceId = characteristic.serviceInstanceId
         ..characteristicUuid =
-            (pb.Uuid()..data = characteristic.characteristicId.data));
+            (pb.Uuid()..data = characteristic.characteristicId.data)
+        ..characteristicInstanceId = characteristic.characteristicInstanceId);
 
     return args;
   }
 
   @override
   pb.WriteCharacteristicRequest createWriteCharacteristicRequest(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
     List<int> value,
   ) {
     final args = pb.WriteCharacteristicRequest()
       ..characteristic = (pb.CharacteristicAddress()
         ..deviceId = characteristic.deviceId
         ..serviceUuid = (pb.Uuid()..data = characteristic.serviceId.data)
+        ..serviceInstanceId = characteristic.serviceInstanceId
         ..characteristicUuid =
-            (pb.Uuid()..data = characteristic.characteristicId.data))
+            (pb.Uuid()..data = characteristic.characteristicId.data)
+        ..characteristicInstanceId = characteristic.characteristicInstanceId)
       ..value = value;
 
     return args;
@@ -137,28 +143,32 @@ class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
 
   @override
   pb.NotifyCharacteristicRequest createNotifyCharacteristicRequest(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
   ) {
     final args = pb.NotifyCharacteristicRequest()
       ..characteristic = (pb.CharacteristicAddress()
         ..deviceId = characteristic.deviceId
         ..serviceUuid = (pb.Uuid()..data = characteristic.serviceId.data)
+        ..serviceInstanceId = characteristic.serviceInstanceId
         ..characteristicUuid =
-            (pb.Uuid()..data = characteristic.characteristicId.data));
+            (pb.Uuid()..data = characteristic.characteristicId.data)
+        ..characteristicInstanceId = characteristic.characteristicInstanceId);
 
     return args;
   }
 
   @override
   pb.NotifyNoMoreCharacteristicRequest createNotifyNoMoreCharacteristicRequest(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
   ) {
     final args = pb.NotifyNoMoreCharacteristicRequest()
       ..characteristic = (pb.CharacteristicAddress()
         ..deviceId = characteristic.deviceId
         ..serviceUuid = (pb.Uuid()..data = characteristic.serviceId.data)
+        ..serviceInstanceId = characteristic.serviceInstanceId
         ..characteristicUuid =
-            (pb.Uuid()..data = characteristic.characteristicId.data));
+            (pb.Uuid()..data = characteristic.characteristicId.data)
+        ..characteristicInstanceId = characteristic.characteristicInstanceId);
 
     return args;
   }
@@ -226,6 +236,12 @@ class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
   @override
   pb.DiscoverServicesRequest createDiscoverServicesRequest(String deviceId) {
     final args = pb.DiscoverServicesRequest()..deviceId = deviceId;
+    return args;
+  }
+
+  @override
+  pb.ReadRssiRequest createReadRssiRequest(String deviceId) {
+    final args = pb.ReadRssiRequest()..deviceId = deviceId;
     return args;
   }
 }
