@@ -115,28 +115,26 @@ void main() {
       const deviceId = '123';
       final serviceUuid = Uuid.parse('FEFF');
       final charUuid = Uuid.parse('FEEF');
-      QualifiedCharacteristic characteristic;
+      CharacteristicInstance characteristic;
 
       setUp(() {
-        characteristic = QualifiedCharacteristic(
+        characteristic = CharacteristicInstance(
           characteristicId: charUuid,
+          characteristicInstanceId: "101",
           serviceId: serviceUuid,
+          serviceInstanceId: "11",
           deviceId: deviceId,
         );
 
         result = _sut.createReadCharacteristicRequest(characteristic);
       });
 
-      test('It converts device Id ', () {
+      test('It converts the ids', () {
         expect(result.characteristic.deviceId, deviceId);
-      });
-
-      test('It converts service Uuid', () {
         expect(result.characteristic.serviceUuid.data, [254, 255]);
-      });
-
-      test('It converts char Uuid', () {
+        expect(result.characteristic.serviceInstanceId, "11");
         expect(result.characteristic.characteristicUuid.data, [254, 239]);
+        expect(result.characteristic.characteristicInstanceId, "101");
       });
     });
 
@@ -145,30 +143,28 @@ void main() {
       const deviceId = '123';
       final serviceUuid = Uuid.parse('FEFF');
       final charUuid = Uuid.parse('FEEF');
-      QualifiedCharacteristic characteristic;
+      CharacteristicInstance characteristic;
 
       const value = [0, 1];
 
       setUp(() {
-        characteristic = QualifiedCharacteristic(
+        characteristic = CharacteristicInstance(
           characteristicId: charUuid,
+          characteristicInstanceId: "101",
           serviceId: serviceUuid,
+          serviceInstanceId: "11",
           deviceId: deviceId,
         );
 
         result = _sut.createWriteCharacteristicRequest(characteristic, value);
       });
 
-      test('It converts device Id ', () {
+      test('It converts the ids', () {
         expect(result.characteristic.deviceId, deviceId);
-      });
-
-      test('It converts service Uuid', () {
         expect(result.characteristic.serviceUuid.data, [254, 255]);
-      });
-
-      test('It converts char Uuid', () {
+        expect(result.characteristic.serviceInstanceId, "11");
         expect(result.characteristic.characteristicUuid.data, [254, 239]);
+        expect(result.characteristic.characteristicInstanceId, "101");
       });
 
       test('It converts value', () {
@@ -181,28 +177,26 @@ void main() {
       const deviceId = '123';
       final serviceUuid = Uuid.parse('FEFF');
       final charUuid = Uuid.parse('FEEF');
-      QualifiedCharacteristic characteristic;
+      CharacteristicInstance characteristic;
 
       setUp(() {
-        characteristic = QualifiedCharacteristic(
+        characteristic = CharacteristicInstance(
           characteristicId: charUuid,
+          characteristicInstanceId: "101",
           serviceId: serviceUuid,
+          serviceInstanceId: "11",
           deviceId: deviceId,
         );
 
         result = _sut.createNotifyCharacteristicRequest(characteristic);
       });
 
-      test('It converts device Id ', () {
+      test('It converts device ids', () {
         expect(result.characteristic.deviceId, deviceId);
-      });
-
-      test('It converts service Uuid', () {
         expect(result.characteristic.serviceUuid.data, [254, 255]);
-      });
-
-      test('It converts char Uuid', () {
+        expect(result.characteristic.serviceInstanceId, "11");
         expect(result.characteristic.characteristicUuid.data, [254, 239]);
+        expect(result.characteristic.characteristicInstanceId, "101");
       });
     });
 
@@ -365,28 +359,26 @@ void main() {
       const deviceId = '123';
       final serviceUuid = Uuid.parse('FEFF');
       final charUuid = Uuid.parse('FEEF');
-      QualifiedCharacteristic characteristic;
+      CharacteristicInstance characteristic;
       late pb.NotifyNoMoreCharacteristicRequest result;
 
       setUp(() {
-        characteristic = QualifiedCharacteristic(
+        characteristic = CharacteristicInstance(
           characteristicId: charUuid,
+          characteristicInstanceId: "101",
           serviceId: serviceUuid,
+          serviceInstanceId: "11",
           deviceId: deviceId,
         );
         result = _sut.createNotifyNoMoreCharacteristicRequest(characteristic);
       });
 
-      test('It converts deviceId', () {
+      test('It converts the ids', () {
         expect(result.characteristic.deviceId, deviceId);
-      });
-
-      test('It converts serviceUuid', () {
         expect(result.characteristic.serviceUuid.data, [254, 255]);
-      });
-
-      test('It converts characteristicUuid', () {
+        expect(result.characteristic.serviceInstanceId, "11");
         expect(result.characteristic.characteristicUuid.data, [254, 239]);
+        expect(result.characteristic.characteristicInstanceId, "101");
       });
     });
 
