@@ -1,14 +1,15 @@
-// Mocks generated by Mockito 5.4.2 from annotations
+// Mocks generated by Mockito 5.4.4 from annotations
 // in reactive_ble_mobile/test/reactive_ble_platform_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i9;
+import 'dart:async' as _i10;
 
 import 'package:flutter/src/services/binary_messenger.dart' as _i5;
 import 'package:flutter/src/services/message_codec.dart' as _i4;
-import 'package:flutter/src/services/platform_channel.dart' as _i8;
+import 'package:flutter/src/services/platform_channel.dart' as _i9;
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:mockito/src/dummies.dart' as _i8;
 import 'package:reactive_ble_mobile/src/converter/args_to_protubuf_converter.dart'
     as _i6;
 import 'package:reactive_ble_mobile/src/converter/protobuf_converter.dart'
@@ -21,6 +22,8 @@ import 'package:reactive_ble_platform_interface/reactive_ble_platform_interface.
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -182,9 +185,20 @@ class _FakeDiscoverServicesRequest_13 extends _i1.SmartFake
         );
 }
 
-class _FakeDeviceAssociationInfo_14 extends _i1.SmartFake
+class _FakeReadRssiRequest_14 extends _i1.SmartFake
+    implements _i2.ReadRssiRequest {
+  _FakeReadRssiRequest_14(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeDeviceAssociationInfo_15 extends _i1.SmartFake
     implements _i3.DeviceAssociationInfo {
-  _FakeDeviceAssociationInfo_14(
+  _FakeDeviceAssociationInfo_15(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -193,8 +207,8 @@ class _FakeDeviceAssociationInfo_14 extends _i1.SmartFake
         );
 }
 
-class _FakeScanResult_15 extends _i1.SmartFake implements _i3.ScanResult {
-  _FakeScanResult_15(
+class _FakeScanResult_16 extends _i1.SmartFake implements _i3.ScanResult {
+  _FakeScanResult_16(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -203,9 +217,9 @@ class _FakeScanResult_15 extends _i1.SmartFake implements _i3.ScanResult {
         );
 }
 
-class _FakeConnectionStateUpdate_16 extends _i1.SmartFake
+class _FakeConnectionStateUpdate_17 extends _i1.SmartFake
     implements _i3.ConnectionStateUpdate {
-  _FakeConnectionStateUpdate_16(
+  _FakeConnectionStateUpdate_17(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -214,9 +228,9 @@ class _FakeConnectionStateUpdate_16 extends _i1.SmartFake
         );
 }
 
-class _FakeResult_17<Value, Failure> extends _i1.SmartFake
+class _FakeResult_18<Value, Failure> extends _i1.SmartFake
     implements _i3.Result<Value, Failure> {
-  _FakeResult_17(
+  _FakeResult_18(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -225,9 +239,9 @@ class _FakeResult_17<Value, Failure> extends _i1.SmartFake
         );
 }
 
-class _FakeCharacteristicValue_18 extends _i1.SmartFake
+class _FakeCharacteristicValue_19 extends _i1.SmartFake
     implements _i3.CharacteristicValue {
-  _FakeCharacteristicValue_18(
+  _FakeCharacteristicValue_19(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -236,9 +250,9 @@ class _FakeCharacteristicValue_18 extends _i1.SmartFake
         );
 }
 
-class _FakeWriteCharacteristicInfo_19 extends _i1.SmartFake
+class _FakeWriteCharacteristicInfo_20 extends _i1.SmartFake
     implements _i3.WriteCharacteristicInfo {
-  _FakeWriteCharacteristicInfo_19(
+  _FakeWriteCharacteristicInfo_20(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -247,9 +261,9 @@ class _FakeWriteCharacteristicInfo_19 extends _i1.SmartFake
         );
 }
 
-class _FakeConnectionPriorityInfo_20 extends _i1.SmartFake
+class _FakeConnectionPriorityInfo_21 extends _i1.SmartFake
     implements _i3.ConnectionPriorityInfo {
-  _FakeConnectionPriorityInfo_20(
+  _FakeConnectionPriorityInfo_21(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -258,8 +272,8 @@ class _FakeConnectionPriorityInfo_20 extends _i1.SmartFake
         );
 }
 
-class _FakeMethodCodec_21 extends _i1.SmartFake implements _i4.MethodCodec {
-  _FakeMethodCodec_21(
+class _FakeMethodCodec_22 extends _i1.SmartFake implements _i4.MethodCodec {
+  _FakeMethodCodec_22(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -268,9 +282,9 @@ class _FakeMethodCodec_21 extends _i1.SmartFake implements _i4.MethodCodec {
         );
 }
 
-class _FakeBinaryMessenger_22 extends _i1.SmartFake
+class _FakeBinaryMessenger_23 extends _i1.SmartFake
     implements _i5.BinaryMessenger {
-  _FakeBinaryMessenger_22(
+  _FakeBinaryMessenger_23(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -303,6 +317,7 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.EstablishBondingRequest);
+
   @override
   _i2.ConnectToDeviceRequest createConnectToDeviceArgs(
     String? id,
@@ -330,6 +345,7 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.ConnectToDeviceRequest);
+
   @override
   _i2.GetDeviceNameRequest createGetDeviceNameArgs(String? id) =>
       (super.noSuchMethod(
@@ -345,6 +361,7 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.GetDeviceNameRequest);
+
   @override
   _i2.DisconnectFromDeviceRequest createDisconnectDeviceArgs(
           String? deviceId) =>
@@ -361,9 +378,10 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.DisconnectFromDeviceRequest);
+
   @override
   _i2.ReadCharacteristicRequest createReadCharacteristicRequest(
-          _i3.QualifiedCharacteristic? characteristic) =>
+          _i3.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #createReadCharacteristicRequest,
@@ -377,9 +395,10 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.ReadCharacteristicRequest);
+
   @override
   _i2.WriteCharacteristicRequest createWriteCharacteristicRequest(
-    _i3.QualifiedCharacteristic? characteristic,
+    _i3.CharacteristicInstance? characteristic,
     List<int>? value,
   ) =>
       (super.noSuchMethod(
@@ -401,9 +420,10 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.WriteCharacteristicRequest);
+
   @override
   _i2.NotifyCharacteristicRequest createNotifyCharacteristicRequest(
-          _i3.QualifiedCharacteristic? characteristic) =>
+          _i3.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #createNotifyCharacteristicRequest,
@@ -417,9 +437,10 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.NotifyCharacteristicRequest);
+
   @override
   _i2.NotifyNoMoreCharacteristicRequest createNotifyNoMoreCharacteristicRequest(
-          _i3.QualifiedCharacteristic? characteristic) =>
+          _i3.CharacteristicInstance? characteristic) =>
       (super.noSuchMethod(
         Invocation.method(
           #createNotifyNoMoreCharacteristicRequest,
@@ -433,6 +454,7 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.NotifyNoMoreCharacteristicRequest);
+
   @override
   _i2.NegotiateMtuRequest createNegotiateMtuRequest(
     String? deviceId,
@@ -457,6 +479,7 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.NegotiateMtuRequest);
+
   @override
   _i2.ChangeConnectionPriorityRequest createChangeConnectionPrioRequest(
     String? deviceId,
@@ -481,6 +504,7 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.ChangeConnectionPriorityRequest);
+
   @override
   _i2.LaunchCompanionRequest createLaunchCompanionWorkflowRequest({
     required String? deviceNamePattern,
@@ -510,6 +534,7 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.LaunchCompanionRequest);
+
   @override
   _i2.ScanForDevicesRequest createScanForDevicesRequest({
     required List<_i3.Uuid>? withServices,
@@ -539,6 +564,7 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.ScanForDevicesRequest);
+
   @override
   _i2.ClearGattCacheRequest createClearGattCacheRequest(String? deviceId) =>
       (super.noSuchMethod(
@@ -554,6 +580,7 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.ClearGattCacheRequest);
+
   @override
   _i2.DiscoverServicesRequest createDiscoverServicesRequest(String? deviceId) =>
       (super.noSuchMethod(
@@ -569,6 +596,22 @@ class MockArgsToProtobufConverter extends _i1.Mock
           ),
         ),
       ) as _i2.DiscoverServicesRequest);
+
+  @override
+  _i2.ReadRssiRequest createReadRssiRequest(String? deviceId) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #createReadRssiRequest,
+          [deviceId],
+        ),
+        returnValue: _FakeReadRssiRequest_14(
+          this,
+          Invocation.method(
+            #createReadRssiRequest,
+            [deviceId],
+          ),
+        ),
+      ) as _i2.ReadRssiRequest);
 }
 
 /// A class which mocks [ProtobufConverter].
@@ -587,6 +630,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
         ),
         returnValue: _i3.BondingStatus.none,
       ) as _i3.BondingStatus);
+
   @override
   _i3.BleStatus bleStatusFrom(List<int>? data) => (super.noSuchMethod(
         Invocation.method(
@@ -595,6 +639,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
         ),
         returnValue: _i3.BleStatus.unknown,
       ) as _i3.BleStatus);
+
   @override
   _i3.DeviceAssociationInfo associationInfoFrom(List<int>? data) =>
       (super.noSuchMethod(
@@ -602,7 +647,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           #associationInfoFrom,
           [data],
         ),
-        returnValue: _FakeDeviceAssociationInfo_14(
+        returnValue: _FakeDeviceAssociationInfo_15(
           this,
           Invocation.method(
             #associationInfoFrom,
@@ -610,13 +655,14 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           ),
         ),
       ) as _i3.DeviceAssociationInfo);
+
   @override
   _i3.ScanResult scanResultFrom(List<int>? data) => (super.noSuchMethod(
         Invocation.method(
           #scanResultFrom,
           [data],
         ),
-        returnValue: _FakeScanResult_15(
+        returnValue: _FakeScanResult_16(
           this,
           Invocation.method(
             #scanResultFrom,
@@ -624,14 +670,22 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           ),
         ),
       ) as _i3.ScanResult);
+
   @override
   String deviceNameFrom(List<int>? data) => (super.noSuchMethod(
         Invocation.method(
           #deviceNameFrom,
           [data],
         ),
-        returnValue: '',
+        returnValue: _i8.dummyValue<String>(
+          this,
+          Invocation.method(
+            #deviceNameFrom,
+            [data],
+          ),
+        ),
       ) as String);
+
   @override
   _i3.ConnectionStateUpdate connectionStateUpdateFrom(List<int>? data) =>
       (super.noSuchMethod(
@@ -639,7 +693,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           #connectionStateUpdateFrom,
           [data],
         ),
-        returnValue: _FakeConnectionStateUpdate_16(
+        returnValue: _FakeConnectionStateUpdate_17(
           this,
           Invocation.method(
             #connectionStateUpdateFrom,
@@ -647,6 +701,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           ),
         ),
       ) as _i3.ConnectionStateUpdate);
+
   @override
   _i3.Result<_i3.Unit, _i3.GenericFailure<_i3.ClearGattCacheError>?>
       clearGattCacheResultFrom(List<int>? data) => (super.noSuchMethod(
@@ -654,7 +709,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
               #clearGattCacheResultFrom,
               [data],
             ),
-            returnValue: _FakeResult_17<_i3.Unit,
+            returnValue: _FakeResult_18<_i3.Unit,
                 _i3.GenericFailure<_i3.ClearGattCacheError>?>(
               this,
               Invocation.method(
@@ -664,6 +719,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
             ),
           ) as _i3
               .Result<_i3.Unit, _i3.GenericFailure<_i3.ClearGattCacheError>?>);
+
   @override
   _i3.CharacteristicValue characteristicValueFrom(List<int>? data) =>
       (super.noSuchMethod(
@@ -671,7 +727,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           #characteristicValueFrom,
           [data],
         ),
-        returnValue: _FakeCharacteristicValue_18(
+        returnValue: _FakeCharacteristicValue_19(
           this,
           Invocation.method(
             #characteristicValueFrom,
@@ -679,6 +735,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           ),
         ),
       ) as _i3.CharacteristicValue);
+
   @override
   _i3.WriteCharacteristicInfo writeCharacteristicInfoFrom(List<int>? data) =>
       (super.noSuchMethod(
@@ -686,7 +743,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           #writeCharacteristicInfoFrom,
           [data],
         ),
-        returnValue: _FakeWriteCharacteristicInfo_19(
+        returnValue: _FakeWriteCharacteristicInfo_20(
           this,
           Invocation.method(
             #writeCharacteristicInfoFrom,
@@ -694,6 +751,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           ),
         ),
       ) as _i3.WriteCharacteristicInfo);
+
   @override
   _i3.ConnectionPriorityInfo connectionPriorityInfoFrom(List<int>? data) =>
       (super.noSuchMethod(
@@ -701,7 +759,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           #connectionPriorityInfoFrom,
           [data],
         ),
-        returnValue: _FakeConnectionPriorityInfo_20(
+        returnValue: _FakeConnectionPriorityInfo_21(
           this,
           Invocation.method(
             #connectionPriorityInfoFrom,
@@ -709,6 +767,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
           ),
         ),
       ) as _i3.ConnectionPriorityInfo);
+
   @override
   int mtuSizeFrom(List<int>? data) => (super.noSuchMethod(
         Invocation.method(
@@ -717,6 +776,7 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
         ),
         returnValue: 0,
       ) as int);
+
   @override
   List<_i3.DiscoveredService> discoveredServicesFrom(List<int>? data) =>
       (super.noSuchMethod(
@@ -726,12 +786,21 @@ class MockProtobufConverter extends _i1.Mock implements _i7.ProtobufConverter {
         ),
         returnValue: <_i3.DiscoveredService>[],
       ) as List<_i3.DiscoveredService>);
+
+  @override
+  int readRssiResultFrom(List<int>? data) => (super.noSuchMethod(
+        Invocation.method(
+          #readRssiResultFrom,
+          [data],
+        ),
+        returnValue: 0,
+      ) as int);
 }
 
 /// A class which mocks [MethodChannel].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockMethodChannel extends _i1.Mock implements _i8.MethodChannel {
+class MockMethodChannel extends _i1.Mock implements _i9.MethodChannel {
   MockMethodChannel() {
     _i1.throwOnMissingStub(this);
   }
@@ -739,26 +808,32 @@ class MockMethodChannel extends _i1.Mock implements _i8.MethodChannel {
   @override
   String get name => (super.noSuchMethod(
         Invocation.getter(#name),
-        returnValue: '',
+        returnValue: _i8.dummyValue<String>(
+          this,
+          Invocation.getter(#name),
+        ),
       ) as String);
+
   @override
   _i4.MethodCodec get codec => (super.noSuchMethod(
         Invocation.getter(#codec),
-        returnValue: _FakeMethodCodec_21(
+        returnValue: _FakeMethodCodec_22(
           this,
           Invocation.getter(#codec),
         ),
       ) as _i4.MethodCodec);
+
   @override
   _i5.BinaryMessenger get binaryMessenger => (super.noSuchMethod(
         Invocation.getter(#binaryMessenger),
-        returnValue: _FakeBinaryMessenger_22(
+        returnValue: _FakeBinaryMessenger_23(
           this,
           Invocation.getter(#binaryMessenger),
         ),
       ) as _i5.BinaryMessenger);
+
   @override
-  _i9.Future<T?> invokeMethod<T>(
+  _i10.Future<T?> invokeMethod<T>(
     String? method, [
     dynamic arguments,
   ]) =>
@@ -770,10 +845,11 @@ class MockMethodChannel extends _i1.Mock implements _i8.MethodChannel {
             arguments,
           ],
         ),
-        returnValue: _i9.Future<T?>.value(),
-      ) as _i9.Future<T?>);
+        returnValue: _i10.Future<T?>.value(),
+      ) as _i10.Future<T?>);
+
   @override
-  _i9.Future<List<T>?> invokeListMethod<T>(
+  _i10.Future<List<T>?> invokeListMethod<T>(
     String? method, [
     dynamic arguments,
   ]) =>
@@ -785,10 +861,11 @@ class MockMethodChannel extends _i1.Mock implements _i8.MethodChannel {
             arguments,
           ],
         ),
-        returnValue: _i9.Future<List<T>?>.value(),
-      ) as _i9.Future<List<T>?>);
+        returnValue: _i10.Future<List<T>?>.value(),
+      ) as _i10.Future<List<T>?>);
+
   @override
-  _i9.Future<Map<K, V>?> invokeMapMethod<K, V>(
+  _i10.Future<Map<K, V>?> invokeMapMethod<K, V>(
     String? method, [
     dynamic arguments,
   ]) =>
@@ -800,11 +877,12 @@ class MockMethodChannel extends _i1.Mock implements _i8.MethodChannel {
             arguments,
           ],
         ),
-        returnValue: _i9.Future<Map<K, V>?>.value(),
-      ) as _i9.Future<Map<K, V>?>);
+        returnValue: _i10.Future<Map<K, V>?>.value(),
+      ) as _i10.Future<Map<K, V>?>);
+
   @override
   void setMethodCallHandler(
-          _i9.Future<dynamic> Function(_i4.MethodCall)? handler) =>
+          _i10.Future<dynamic> Function(_i4.MethodCall)? handler) =>
       super.noSuchMethod(
         Invocation.method(
           #setMethodCallHandler,
